% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{blablabla}
\alias{blablabla}
\title{Print messages}
\usage{
blablabla(msg, silent = FALSE)
}
\arguments{
\item{msg}{[\code{character(1)}]\cr the message.}

\item{silent}{[\code{logical(1)}]\cr should all messages except errors be
suppressed (\code{TRUE}), or should all messages be printed (\code{FALSE},
default)?}
}
\description{
This is a wrapper around \code{\link{message}}.
}
