% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{catalog}
\alias{catalog}
\title{Build a look-up-table (catalog)}
\usage{
catalog(path = ".", abbreviateBy = NULL, type = NULL,
  recursive = FALSE, inclType = FALSE, silent = FALSE)
}
\arguments{
\item{path}{[\code{character(1)}]\cr the path to a folder that contains files
for which you want to build a catalog.}

\item{abbreviateBy}{[\code{function}]\cr to translate the original names to
abbreviations thereof; see Examples.}

\item{type}{[\code{character(1)}]\cr the file type for which should be
filtered.}

\item{recursive}{[\code{logical(1)}]\cr should the listing recurse into sub
directories (\code{TRUE}) or should only the specified directory be
searched (\code{FALSE}, default)?}

\item{inclType}{[\code{logical(1)}]\cr should the file-extension be included
in the recorded original names (\code{TRUE}), or should the file name be
included without extension (\code{FALSE}, default)?}

\item{silent}{[\code{logical(1)}]\cr should all messages but errors be
suppressed (\code{TRUE}), or should all messages be printed (\code{FALSE},
default)?}
}
\value{
A \code{data.frame} with a column of original names and of the
  associated abreviations.
}
\description{
\code{catalog} aims to make working with a wide variety of complex names
simpler. Complex terms such as species, territorial units, expriment names,
... are related to systematic abbreviations thereof.
}
\details{
Often data-files come with complicated names, characterising in
  large detail the content. For a quick workflow, this is often a hindrance
  and not required, because a shorter yet systematic name can also
  sufficiently reference files. Typical examples would be: \itemize{ \item
  three or four-letter code for species (apo_fla, myo_gla, ...). \item
  two-letter NUTS code of territorial units (EE, DE, ...). \item a short name
  of some repeated experiment, originally labelled by
  date_region_researcher_treatment_etc (S11, S12, ...). \item ... } If you
  want to work with a relatively large amount of such files on our hard
  drive, it would be desirable to have some sort of index available, which
  makes it possible to readily access everything without having to load it to
  the global environment beforehand. This function tries to help here, by
  iteratively going through all directories which are specified in
  \code{path}, registering the files (incl. \code{path}) of the specified
  \code{type} and deriving a name based on the function defined in
  \code{abbreviateBy}. These information are stored in the \code{data.frame}
  \code{myCatalog} in the gloabl environment so that the objects in
  \code{path} can be quickly loaded by referring to their abbreviations.

  The recorded original names are internally saved as character vector with
  the name \code{files} and \code{i} is used as iterator through its
  elements. The only limitation for this function hence is that the
  abbreviation must be some derivative of the original filenames, employing
  character manipulation functions such as \code{\link{substr}},
  \code{\link{strsplit}}, \code{\link{toupper}}, \code{\link{paste0}} or any
  other.
}
\examples{
\dontrun{

# build a catalog for all species for which there are maps in the 'Atlas of
# European mammals'

# 1. create character variable which corresponds to your file names
x <- "Vulpes vulpes.svg"

# 2. combine with character manipulation functions until the pattern matches
strsplit(x, ' ')
unlist(strsplit(x, ' '))
substr(unlist(strsplit(x, ' ')), 1, 3)
paste(substr(unlist(strsplit(x, ' ')), 1, 3), collapse = '_')
tolower(paste(substr(unlist(strsplit(x, ' ')), 1, 3), collapse = '_'))

# 3. turn this into a function and use as value in abbreviateBy.
abbr_species <- function(x){
  tolower(paste(substr(unlist(strsplit(x, ' ')), 1, 3), collapse = '_'))
}

require(magrittr)
catalog(path = rtPaths$emma$local, type = 'svg', inclType = TRUE,
    abbreviateBy = abbr_species) \%>\%
  subset(abbr \%in\% c("apo_agr", "apo_fla", "vul_vul"))
}
}
