% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{doyToDate}
\alias{doyToDate}
\title{Translate 'day of the year' to a date}
\usage{
doyToDate(year, doy)
}
\arguments{
\item{year}{[\code{integerish(1)}]\cr a year value as part of the date.}

\item{doy}{[\code{integerish(1)}]\cr the day of the year for which you want
the date}
}
\value{
An object of class \code{\link{POSIXct}}.
}
\description{
Translate 'day of the year' to a date
}
\examples{
aDate <- doyToDate(year = 2000, doy = 055)
}
