% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometries.R
\name{geomPoint}
\alias{geomPoint}
\title{Create a point geometry}
\usage{
geomPoint(anchor = NULL, window = NULL, template = NULL,
  vertices = NULL, show = FALSE, ...)
}
\arguments{
\item{anchor}{[\code{data.frame(1)}]\cr Object to derive the \code{geom}
from. It must include column names \code{x}, \code{y} and optinal variables
such as \code{id}; see Examples.}

\item{window}{[\code{data.frame(1)}]\cr in case the reference window deviates
from the bounding box of \code{anchor} (minimum and maximum values),
specify this here.}

\item{template}{[\code{RasterLayer(1)} | \code{matrix(1)}]\cr Gridded object
that serves as template to sketch the geometry.}

\item{vertices}{[\code{integer(1)}]\cr number of vertices.}

\item{show}{[\code{logical(1)}]\cr in case \code{template} is set, should the
points be plotted (\code{TRUE}) or should it not be plotted (\code{FALSE},
default)?}

\item{...}{[various]\cr graphical parameter, in case \code{show = TRUE}; see
\code{\link{gpar}}.}
}
\value{
An invisible \code{geom}.
}
\description{
Create a point geometry
}
\examples{
# create points programmatically
somePoints <- data.frame(id = 1:8, X = c(5190599, 5222810, 5041066, 5234735,
                         5326537, 5027609, 5281527, 5189955), Y = c(3977612,
                         4060164, 3997230, 4117856, 4028167, 3971119, 4118207,
                         4062838))
(pointsGeom <- geomPoint(anchor = somePoints, col = "darkorange", show = TRUE))

\dontrun{

input <- rtData$continuous

# create points interactively
myPoints <- geomPoint(template = input, vertices = 5, show = TRUE, col = "deeppink")
anExtent <- geomRectangle(myPoints, show = TRUE, col = "green")
}
}
\seealso{
Tools to modify geometries: \code{\link{gRotate}},
  \code{\link{gScale}}, \code{\link{gGroup}}\cr Tools to transform geometries
  to other classes: \code{\link{gToSp}},  \code{gToSf},
  \code{\link{gToRaster}}, \code{\link{gFrom}}

Other shapes: \code{\link{geomPolygon}},
  \code{\link{geomRand}}
}
