% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometries.R
\name{geomRand}
\alias{geomRand}
\title{Create a geometry randomly}
\usage{
geomRand(type = "point", template = NULL, vertices = NULL, show = FALSE)
}
\arguments{
\item{type}{[\code{character(1)}]\cr Either one of the three main feature
types \code{"point"}, \code{"line"} or \code{"polygon"}, or more
specifically one of their subtypes, e.g. \code{"hexagon"}.}

\item{template}{[\code{RasterLayer(1)} | \code{matrix(1)}]\cr Gridded object
that serves as template to sketch the geometry.}

\item{vertices}{[\code{integersh(1)}]\cr the number of vertices the geometry
should have; only meaningful if \code{type} does not indicate the number of
vertices already.}

\item{show}{[\code{logical(1)}]\cr should the geometry be plotted
(\code{TRUE}) or should it not be plotted (\code{FALSE}, default)? In case
\code{template} is set, it is automatically \code{TRUE}.}
}
\description{
This function creates, depending on the input parameters, a random geometry
}
\examples{
input <- matrix(nrow = 100, ncol = 100, data = 0)

# create a random geometry with four vertices
someGeom <- geomRand(type = "polygon", vertices = 4)

# in case template is given, this serves as source for the window extent
someGeom <- geomRand(template = input, vertices = 5)
}
\seealso{
Other shapes: \code{\link{geomPoint}},
  \code{\link{geomPolygon}}
}
