% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometries.R
\name{geomTiles}
\alias{geomTiles}
\title{Create a regular tiling geometry}
\usage{
geomTiles(window = NULL, cells = NULL, crs = NULL,
  tiling = "rectangular", centroids = FALSE)
}
\arguments{
\item{window}{[\code{data.frame(1)}]\cr the origin (lower left corner) and
the maximum value (upper right corner) of the tiling.}

\item{cells}{[\code{integerish(2)}]\cr number of tiles in \code{x} and
\code{y} dimension.}

\item{crs}{[\code{proj4string(1)}]\cr corrdinate reference system of the
object.}

\item{tiling}{[\code{¢haracter(1)}]\cr pattern of the tiling. Possible
options are \code{"rectangular"} (default), \code{"hexagonal"},
\code{"triangular"}.}

\item{centroids}{[\code{logical(1)}]\cr should the centroids of the tiling be
returned (\code{TRUE}) or should the tiling be returned (\code{FALSE},
default)?}
}
\value{
An invisible \code{geom}.
}
\description{
Create a regular tiling geometry
}
\details{
https://code.env.duke.edu/projects/mget/wiki/SinusoidalMODIS

  origin is in the lower left corner
}
\examples{
# create grid for GFC data
gfcWindow <- data.frame(x = c(-180, 180),
                        y = c(-60, 80))
tiles_gfc <- geomTiles(window = gfcWindow, cells = c(36, 14), crs = projs$longlat)

# create grid for MODIS data
modWindow <- data.frame(x = c(-20015109.354, 21127059.874),
                        y = c(-10007554.677, 11119505.197))
tiles_modis <- geomTiles(window = modWindow, cells = c(37, 19), crs = projs$sinu)

# create grid for the sentinel data
#sntWindow <- data.frame(x = c(),
#                        y = c())
#tiles_sentinel <- geomTiles(window = sntWindow, cells = c(), crs = projs$utm)
}
