% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.R
\docType{methods}
\name{getColumn}
\alias{getColumn}
\alias{getColumn,geom,numeric-method}
\alias{getColumn,geom,character-method}
\alias{getColumn,geom,logical-method}
\title{Get variables or columns (generic)}
\usage{
getColumn(x, column, ...)

\S4method{getColumn}{geom,numeric}(x, column)

\S4method{getColumn}{geom,character}(x, column)

\S4method{getColumn}{geom,logical}(x, column)
}
\arguments{
\item{x}{the object from which to get/extract an element.}

\item{column}{the column to extract.}

\item{...}{other arguments.}
}
\description{
This function is equivalent to the \code{[} function on the
\code{@table} slot of a geom and extracts columns from this table.
}
\section{Methods (by class)}{
\itemize{
\item \code{x = geom,column = numeric}: get a column of the table of a \code{geom} base on a numeric

\item \code{x = geom,column = character}: get a column of the table of a \code{geom} base on a character

\item \code{x = geom,column = logical}: get a column of the table of a \code{geom} base on a logical
}}

