% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.R
\docType{methods}
\name{getCRS}
\alias{getCRS}
\alias{getCRS,geom-method}
\alias{getCRS,Raster-method}
\alias{getCRS,Spatial-method}
\title{Get the coordinate reference system (generic)}
\usage{
getCRS(x, ...)

\S4method{getCRS}{geom}(x)

\S4method{getCRS}{Raster}(x)

\S4method{getCRS}{Spatial}(x)
}
\arguments{
\item{x}{the object from which to derive the coordinate reference system.}

\item{...}{other arguments.}
}
\description{
Get the coordinate reference system (generic)
}
\section{Methods (by class)}{
\itemize{
\item \code{geom}: get the coordinate reference system of a \code{geom}

\item \code{Raster}: get the coordinate reference system of a \code{Raster*} object

\item \code{Spatial}: get the coordinate reference system of a \code{Spatial*} object
}}

\seealso{
\code{\link{setCRS}}
}
