% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.R
\docType{methods}
\name{getExtent}
\alias{getExtent}
\alias{getExtent,geom-method}
\alias{getExtent,Raster-method}
\alias{getExtent,Spatial-method}
\title{Get the extent (generic)}
\usage{
getExtent(x, ...)

\S4method{getExtent}{geom}(x)

\S4method{getExtent}{Raster}(x)

\S4method{getExtent}{Spatial}(x)
}
\arguments{
\item{x}{the object from which to derive the extent.}

\item{...}{other arguments.}
}
\description{
Get the extent (generic)
}
\section{Methods (by class)}{
\itemize{
\item \code{geom}: get the bounding box of a \code{geom}

\item \code{Raster}: get the bounding box of a \code{Raster*} object

\item \code{Spatial}: get the bounding box of a \code{Spatial*} object
}}

