% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.R
\docType{methods}
\name{getRow}
\alias{getRow}
\alias{getRow,geom,numeric-method}
\alias{getRow,geom,logical-method}
\title{Get rows (generic)}
\usage{
getRow(x, row, ...)

\S4method{getRow}{geom,numeric}(x, row)

\S4method{getRow}{geom,logical}(x, row)
}
\arguments{
\item{x}{the object from which to get/extract an element.}

\item{row}{the row to extract.}

\item{...}{other arguments.}
}
\description{
This function is equivalent to the \code{[} function on the
\code{@table} slot of a geom and extracts rows from this table.
}
\section{Methods (by class)}{
\itemize{
\item \code{x = geom,row = numeric}: get a row of a \code{geom} based on a numeric

\item \code{x = geom,row = logical}: get a row of a \code{geom} based on a logical
}}

