% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.R
\docType{methods}
\name{getTable}
\alias{getTable}
\alias{getTable,geom-method}
\title{Get the table (generic)}
\usage{
getTable(x, ...)

\S4method{getTable}{geom}(x)
}
\arguments{
\item{x}{the object from which to derive the attribute table.}

\item{...}{other arguments.}
}
\description{
Get the table (generic)
}
\section{Methods (by class)}{
\itemize{
\item \code{geom}: get the attribute table (including coordinates) of a \code{geom}
}}

