% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.R
\docType{methods}
\name{getWindow}
\alias{getWindow}
\alias{getWindow,geom-method}
\title{Get the reference window (generic)}
\usage{
getWindow(x, ...)

\S4method{getWindow}{geom}(x)
}
\arguments{
\item{x}{the object from which to derive the reference window.}

\item{...}{other arguments.}
}
\description{
Get the reference window (generic)
}
\section{Methods (by class)}{
\itemize{
\item \code{geom}: get the reference window of a \code{geom}
}}

