% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_tools.r
\name{gFrom}
\alias{gFrom}
\title{Derive a \code{geom} from other spatial objects}
\usage{
gFrom(input)
}
\arguments{
\item{input}{[various]\cr the spatial object; currently defined are ojects of
class \code{sp} and \code{sf}.}
}
\value{
a \code{geom} of the type that comes closest to the type of the
  input.
}
\description{
See \code{\link{geom-class}} for details on differences between objects of
class \code{geom} and other spatial classes.
}
\examples{
require(sp)
pol1 <- Polygon(cbind(c(4, 4, 5, 7, 4),c(5, 3, 2, 5, 5)))
pol2 <- Polygon(cbind(c(4.5, 5.5, 5.5, 5, 4.5, 4.5),
                      c(4.5, 4, 3.5, 3, 3, 4)),
                      hole = TRUE)
poly <- Polygons(list(pol1, pol2), "1")
spPoly <- SpatialPolygons(list(poly))
plot(spPoly, col = "goldenrod")

myGeom <- gFrom(input = spPoly)
visualise(geom = myGeom)

}
