% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_tools.r
\name{gRotate}
\alias{gRotate}
\title{Rotate geometries}
\usage{
gRotate(geom, angle, about = c(0, 0), id = NULL)
}
\arguments{
\item{geom}{[\code{geom}]\cr Object of class \code{\link{geom}}.}

\item{angle}{[\code{numeric(1)}]\cr the counter-clockwise angle by which
\code{geom} shall be rotated.}

\item{about}{[\code{numeric(2)}]\cr the point about which \code{geom} shall
be rotated.}

\item{id}{[\code{integerish(.)}]\cr vector of ids that should be rotated.}
}
\value{
Rotated \code{geom}.
}
\description{
Rotate \code{geom}s by a certain angle about a center
}
\examples{
coords <- data.frame(x = c(30, 60, 60, 40, 10, 40, 20),
                     y = c(40, 40, 60, 70, 10, 20, 40),
                     id = c(1, 1, 1, 1, 2, 2, 2))
window <- data.frame(x = c(0, 80),
                     y = c(0, 80))
aGeom <- geomPolygon(anchor = coords, window = window,
                     col = "blue", show = TRUE)

rotatedGeom <- gRotate(geom = aGeom, angle = 90, about = c(40, 40))
visualise(geom = rotatedGeom)

# rotate single objects
rotatedTriangle <- gRotate(geom = aGeom, angle = -180, about = c(30, 40), id = 2)
visualise(geom = rotatedTriangle, col = "goldenrod1")

# rotate different objects about different centers by different angles
rotateMore <- gRotate(geom = aGeom,
                      angle = list(90, -180),
                      about = list(c(40, 40), c(30, 40)))
visualise(geom = rotateMore, col = "deeppink")
}
