% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_tools.r
\name{gToGrob}
\alias{gToGrob}
\title{Transform geometry to grob}
\usage{
gToGrob(geom, theme = NULL, ...)
}
\arguments{
\item{geom}{[\code{geom}]\cr Object of class \code{\link{geom}}.}

\item{theme}{[\code{list(7)}]\cr Visualising options; see
\code{\link{setTheme}} for details.}

\item{...}{instead of providing a \code{theme}, you can also determine
specific graphic parameters (see \code{\link{gpar}}) separately.}
}
\value{
Depending on the provided geometry either a \code{\link{pointsGrob}},
  \code{\link{linesGrob}}, \code{\link{polylineGrob}} or a
  \code{\link{polygonGrob}}.
}
\description{
A \code{\link{grob}} (graphical object) is the grid-package representation of
a \code{geom} and is used for plotting.
}
\examples{
coords <- data.frame(x = c(40, 70, 70, 50),
                     y = c(40, 40, 60, 70),
                     id = 1)
window <- data.frame(x = c(0, 80),
                     y = c(0, 80))
aGeom <- geomPolygon(anchor = coords, window = window, col = "blue")

aGrob <- gToGrob(geom = aGeom)
str(aGrob)
}
