% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_tools.r
\name{gToSp}
\alias{gToSp}
\title{Transform geometry to spatial object}
\usage{
gToSp(geom, crs = NULL)
}
\arguments{
\item{geom}{[\code{geom}]\cr Object of class \code{\link{geom}}.}

\item{crs}{[\code{proj4string(1)}]\cr corrdinate reference system of the
object.}
}
\value{
Depending on the provided geometry either a
  \code{\link{SpatialPointsDataFrame}}, \code{\link{SpatialLinesDataFrame}}
  or \code{\link{SpatialPolygonsDataFrame}} object.
}
\description{
An object of class \code{Spatial*} is the sp-package representation of a
\code{geom}.
}
\examples{
require(magrittr)
somePoints <- data.frame(X = c(5027609, 5190599, 5326537, 5222810,
                         5234735, 5281527, 5189955, 5041066),
                         Y = c(3977612, 3971119, 4028167, 3997230,
                         4060164, 4117856, 4118207, 4062838),
                         id = c(1:8))

pointsGeom <- geomPoints(anchor = somePoints)
polyGeom <- gGroup(geom = pointsGeom, index = c(rep(1, 8))) \%>\%
  geomPolygon()

(spPoints <- gToSp(geom = pointsGeom, crs = projs$laea))
(spPolygon <- gToSp(geom = polyGeom, crs = projs$laea))
}
