% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{load_hdf}
\alias{load_hdf}
\title{Load \code{hdf} files}
\usage{
load_hdf(path, layer)
}
\arguments{
\item{path}{[\code{character(1)}]\cr the local path of the file to load.}

\item{layer}{[\code{character(1)}]\cr the layer name.}
}
\value{
\code{RasterLayer} of the loaded \code{hdf} file
}
\description{
hdf is the short form of \emph{Hierarchical Data Format}, which is a
standardised file format to store scientific data. There are two commonly
used version of this format, hdf4 and hdf5. For now, \code{load_hdf} loads
only files of format hdf5. It is a wrapper of \link[gdalUtils]{gdalinfo}.
}
\details{
This is a helper to \code{\link{loadData}} and is not intended to be used on
its own.
}
\seealso{
Other loaders: \code{\link{load_csv}},
  \code{\link{load_kml}}, \code{\link{load_svg}}
}
