% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{load_svg}
\alias{load_svg}
\title{Load \code{svg} files}
\usage{
load_svg(path, layer)
}
\arguments{
\item{path}{[\code{character(1)}]\cr the local path of the file to load.}

\item{layer}{[\code{character(1)}]\cr the origin of the dataset that is
captured by the file. Recently this is only \code{"emma"}.}
}
\value{
\code{data.frame} of the content of interest depending on
  \code{layer}.
}
\description{
svg is the short form of \emph{Scalable Vector Graphics}. These files are
basically text files interpreted and consequently visualised by any program
that is suitable to do so (such as your web-browser or
\href{www.inkscape.org}{inkscape}). This means that everything that is
visible in the image, is somehow coded into the text. Consequently, in case
the file is systematic and contains information of interest, these can be
turned into computer readable data.
}
\details{
This is a helper to \code{\link{loadData}} and is not intended to be used on
its own.
}
\seealso{
Other loaders: \code{\link{load_csv}},
  \code{\link{load_hdf}}, \code{\link{load_kml}}
}
\concept{loaders}
