% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Import spatial datasets}
\usage{
loadData(files = NULL, dataset = NULL, layer = NULL, localPath = NULL,
  verbose = FALSE, envir = .GlobalEnv)
}
\arguments{
\item{files}{[\code{character(1)} | \code{list(.)}]\cr the file, as it is
called in \code{localPath} or \code{onlinePath}. \code{files} can be a list
to (down)load several files in the same directory/url with one call.
\code{files} also accepts the (subsetted) output of \code{\link{catalog}},
i.e. a \code{data.frame} with the two columns \code{original} and
\code{abbr}.}

\item{dataset}{[\code{character(1)}]\cr character vector of the data dataset
for which \code{files} should be imported; if this is given,
\code{localPath} and \code{onlinePath} are overwritten with the internal
dataset default (see \code{\link{updatePaths}}).}

\item{layer}{[\code{character(1)} | \code{integerish(1)}]\cr name or position
of the layer, other than the \code{\link[rgdal]{readOGR}}-default, that
should be loaded.}

\item{localPath}{[\code{character(1)}]\cr the local path where files are
supposed to be stored (i.e. on your harddisc).}

\item{verbose}{[\code{logical(1)}]\cr should additional information be
printed (\code{TRUE}) or should it be suppressed (\code{FALSE}, default)?}

\item{envir}{[\code{environment(1)}]\cr the environment to which the output
of this function should be saved.}
}
\value{
a (list of) file(s). If there are several files of type
  \code{Raster*}, they are stacked.
}
\description{
Meta function to import, i.e. load or download and unpack spatial datasets
and other files containing spatial data.
}
\details{
\code{importData} checks whether the required files are available in
  \code{localPath}. If nothing is found there, it attempts to download the
  files from \code{onlinePath}. \code{rasterTools} contains a predefined internal
  index, \code{getOption("rtPaths")}, where the local and online paths of
  the supported spatial datasets are stored. Please take a look at it and in
  case an online resource has been altered, you can adapt it here (see
  \code{\link{updatePaths}}).
}
\examples{
\dontrun{

# load a batch of '.csv'-files containing coordinates into the gloabl environment.
loadData(files = c("myLocations1.csv", "myLocations2.csv"),
         localPath = rtPaths$project, verbose = TRUE)

}
}
\seealso{
The specific load operators: \code{\link{load_kml}},
  \code{\link{load_csv}}, \code{\link{load_kml}}, \code{\link{load_hdf}},
  \code{load_tif}, \code{load_dbf}, \code{load_shp}, \code{\link{load_svg}}
}
