% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_generic.R
\name{mAdjacency}
\alias{mAdjacency}
\title{Adjacency of cells}
\usage{
mAdjacency(obj, type = "like", count = "double", layer = NULL)
}
\arguments{
\item{obj}{[\code{RasterLayer(1)}]\cr The object to measure.}

\item{type}{[\code{character(1)}]\cr which type of adjacencies to calculate;
either \code{"like"} adjacencies between cells of the same value,
adjacencies between \code{"paired"} values or the rowsum of the paired
values (\code{"pairedSum"}).}

\item{count}{[\code{character(1)}]\cr the cells that should be counted;
possible values are \code{"single"} to count only cells to the right and
bottom of the focal cell and \code{"double"} to count additionally cells to
the left and top of the focal cell.}

\item{layer}{[\code{character(1)}]\cr in case \code{obj} has several layers,
specify here the layer for which the area of objects shall be calculated
(by default, the first layer).}
}
\value{
a \code{data.frame} of the frequency of adjacencies of the values in
  \code{obj}.
}
\description{
Calculate the cell-adjacency matrix for a raster
}
\details{
In case \code{type = "like"}, only the diagonal of the adjacency
  matrix is returned, in case \code{type = "paired"}, the complete adjacency
  matrix is returned.
}
\examples{
cat <- rtData$categorial
bin <- rBinarise(rtData$continuous, thresh = 40)

# couble count like adjacencies
mAdjacency(obj = cat)

# paired adjacencies
mAdjacency(obj = cat, type = "paired")

# adjacencies with single count
mAdjacency(obj = bin, count = "single")

}
\seealso{
Other generic metrics: \code{\link{mArea}},
  \code{\link{mNumber}}, \code{\link{mPerimeter}},
  \code{\link{mValues}}
}
