% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_derived.R
\docType{data}
\name{mCPA}
\alias{mCPA}
\title{Derived metric: Class proportion of the Landscape}
\format{A list of 3 elements:
\describe{
  \item{a_c}{Area per class: list(operator = "mArea", scale = "class")}
  \item{a_p}{Area of the window: list(operator = "mArea", scale = "window")}
  \item{equation}{"a_c / a_w * 100"}
}}
\usage{
mCPA
}
\description{
Proportion of the landscape window occupied by classes.
}
\keyword{datasets}
