% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_derived.R
\docType{data}
\name{mDE}
\alias{mDE}
\title{Derived metric: Edge density}
\format{A list of 3 elements:
\describe{
  \item{p_p}{Perimeter per patch: list(operator = "mPerimeter", scale = "patch")}
  \item{a_w}{Area of the window: list(operator = "mArea", scale = "window")}
  \item{equation}{"p_p / a_w * 10000"}
}}
\usage{
mDE
}
\description{
Length of edges of patches per hectare of the landscape window.
}
\keyword{datasets}
