% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_derived.R
\docType{data}
\name{mDEc}
\alias{mDEc}
\title{Derived metric: Edge density}
\format{A list of 3 elements:
\describe{
  \item{p_c}{Perimeter per class: list(operator = "mPerimeter", scale = "class")}
  \item{a_w}{Area of the window: list(operator = "mArea", scale = "window")}
  \item{equation}{"p_c / a_w * 10000"}
}}
\usage{
mDEc
}
\description{
Length of edges of classes per hectare of the landscape window.
}
\keyword{datasets}
