% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_derived.R
\docType{data}
\name{mDP}
\alias{mDP}
\title{Derived metric: Patch density}
\format{A list of 3 elements:
\describe{
  \item{n_p}{Number of patches: list(operator = "mNumber", scale = "patch")}
  \item{a_w}{Area of the window: list(operator = "mArea", scale = "window")}
  \item{equation}{"n_p / a_w * 10000"}
}}
\usage{
mDP
}
\description{
Number of patches per hectare of the landscape window.
}
\keyword{datasets}
