% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta_efta}
\alias{meta_efta}
\title{Metadata of the EFTA datasets}
\format{The \code{data.frame} with 4 variables and 43 entries contains
  meta-data for the tree species covered in the European Atlas of Forest Tree
  Species. \describe{ \item{\code{botanical}}{the species name as noted in
  the European Atlas of Forest Tree Species} \item{\code{english}}{the
  english name} \item{\code{rpp}}{a binary variable whether the 'relative
  probability of presence' dataset is available for this species}
  \item{\code{mhs}}{a binary variable whether the 'maximum habitat
  suitability' dataset is available for this species}}}
\usage{
meta_efta
}
\description{
Metadata of the EFTA datasets
}
\keyword{datasets}
