% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta_emma}
\alias{meta_emma}
\title{Metadata of the EMMA datasets}
\format{The \code{data.frame} with 5 variables and 245 entries contains
  meta-data for the mammals covered in the Atlas of European Mammals.
  \describe{ \item{\code{original}}{the species name as noted in the Atlas of
  European Mammals} \item{\code{english}}{the english name}
  \item{\code{author}}{the species author} \item{\code{order}}{the species
  order} \item{\code{family}}{the species family} \item{\code{iucn}}{the
  conservation status according to the IUCN Red List of Threatened Species}}}
\usage{
meta_emma
}
\description{
Metadata of the EMMA datasets
}
\keyword{datasets}
