% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta_esalc}
\alias{meta_esalc}
\title{Metadata of the ESA CCI land-cover datasets}
\format{The \code{data.frame} with 3 variables and 38 entries contains
  meta-data for the classes of land cover in the esalc dataset. \describe{
  \item{\code{NB_LAB}}{integer value of the land-cover class} 
  \item{\code{LCCOwnLabel}}{the long class name}
  \item{\code{colour}}{hexadecimal code of the colour per land-cover class.}}}
\usage{
meta_esalc
}
\description{
Metadata of the ESA CCI land-cover datasets
}
\keyword{datasets}
