% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta_modis}
\alias{meta_modis}
\title{Metadata of the MODIS datasets}
\format{The dataframe with 13 variables and 97 entries contains meta-data for
  a couple of predefined MODIS products. \describe{ \item{\code{product}}{the
  MODIS-product name as defined by NASA} \item{\code{newest_version}}{the
  newest version for the respective product}
  \item{\code{temporal_granularity}}{the number of days between two
  data-files} \item{\code{pixel_size}}{the spatial resolution, i.e. the size
  of one raster cell} \item{\code{sds_layer_name}}{the machine-readable name
  of the layer} \item{\code{description}}{a human-readable description of the
  layer} \item{\code{units}}{the unit in which the data are saved}
  \item{\code{data_type}}{bit number which determines the length of the
  values} \item{\code{fill_min}}{minimum fill value}
  \item{\code{fill_max}}{maximum fill value} \item{\code{valid_min}}{the
  minimum valid value} \item{\code{valid_max}}{the maximum valid value}
  \item{\code{correction_factor}}{a correction factor}}}
\usage{
meta_modis
}
\description{
Metadata of the MODIS datasets
}
\keyword{datasets}
