% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_derived.R
\docType{data}
\name{mLPI}
\alias{mLPI}
\title{Derived metric: Largest Patch Index}
\format{A list of 3 elements:
\describe{
  \item{a_w}{Area of the window: list(operator = "mArea", scale = "window")}
  \item{a_p}{Area per patch: list(operator = "mArea", scale = "patch")}
  \item{equation}{"max(a_p) / a_w * 100"}
}}
\usage{
mLPI
}
\description{
Proportion of the landscape window occupied by the largest patch.
}
\keyword{datasets}
