% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.operators.R
\name{mMetric}
\alias{mMetric}
\title{Metric distribution}
\usage{
mMetric(obj, metric)
}
\arguments{
\item{obj}{[\code{Raster*(1)}]\cr The object to measure.}

\item{metric}{[\code{character(.)}]\cr the metric(s) that shall be assessed.}
}
\description{
Metric distribution
}
