% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.operators.R
\name{mNumber}
\alias{mNumber}
\title{Number of objects}
\usage{
mNumber(obj, scale = "patch", layer = NULL)
}
\arguments{
\item{obj}{[\code{RasterLayer(1)}]\cr The object to measure.}

\item{scale}{[\code{character(1)}]\cr scale at which the area of objects
should be calculated; possible values are \code{"patch"} and \code{"class"}.}

\item{layer}{[\code{character(1)}]\cr in case \code{obj} has several layers,
specify here the layer for which the area of objects shall be calculated
(by default, the first layer).}
}
\value{
For \code{scale = "class"} the number of unique values (classes) in
  the raster. For \code{scale = "patch"} the number of objects per distinc
  value (i.e. the number of patches per class).
}
\description{
Count the number of objects in a raster
}
\examples{
cat <- rtData$categorical
bin <- rBinarise(rtData$continuous, thresh = 40)

# the number ...
# ... per landuse type
mNumber(obj = cat, scale = "class")

# ... of patches per landuse type
mNumber(obj = cat, scale = "patch")

# ...  of certain values
require(magrittr)
rBinarise(obj = cat, match = c(41, 44, 47)) \%>\%
  mNumber(scale = "patch", layer = "values_binarised")

# ... from a binary raster
mNumber(obj = bin, scale = "patch")
mNumber(obj = bin, scale = "class")

}
\seealso{
Other generic metrics: \code{\link{mAdjacency}},
  \code{\link{mArea}}, \code{\link{mPerimeter}},
  \code{\link{mValues}}
}
