% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_derived.R
\docType{data}
\name{mPAR}
\alias{mPAR}
\title{Derived metric: Perimeter-Area ratio}
\format{A list of 3 elements:
\describe{
  \item{p_p}{Perimeter per patch: list(operator = "mPerimeter", scale = "patch")}
  \item{a_p}{Area per patch: list(operator = "mArea", scale = "patch")}
  \item{equation}{"p_p / a_p * 10000"}
}}
\usage{
mPAR
}
\description{
Ratio of the perimeter per area of each patch.
}
\keyword{datasets}
