% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.operators.R
\name{mPerimeter}
\alias{mPerimeter}
\title{Edge length}
\usage{
mPerimeter(obj, scale = "patch", unit = "cells", layer = NULL)
}
\arguments{
\item{obj}{[\code{Raster*(1)}]\cr The object to measure.}

\item{scale}{[\code{character(1)}]\cr scale at which the area of objects
should be calculated; possible values are \code{"patch"} and
\code{"class"}.}

\item{unit}{[\code{character(1)}]\cr the unit the output should have. With
\code{"map"} the result will be in the respective map unit and with
\code{"cells"} (default) it will be the multiple of the raster cell
dimension.}

\item{layer}{[\code{character(1)}]\cr in case \code{obj} has several layers,
specify here the layer for which the area of objects shall be calculated
(by default, the first layer).}
}
\value{
For \code{scale = "class"} the edge length of each unique value
  (class) in the raster, for \code{scale = "patch"} the edge length of
  distinct objects per distinct values (i.e. the area of patches per class).
}
\description{
Calculate the length of edges/boundaries between objects in a spatial raster.
}
\examples{
cat <- rtData$categorical
bin <- rBinarise(rtData$continuous, thresh = 40)

# the perimeter ...
# ... per landcover type
mPerimeter(obj = cat, scale = "class")

# ... of patches per landcover typ
mPerimeter(obj = cat, scale = "patch")

# ...  of certain values; from a binary raster, patches are 
# automatically determined
require(magrittr)
rBinarise(obj = cat, match = c(41, 44, 47)) \%>\%
  mPerimeter(scale = "patch", layer = "values_binarised")

mPerimeter(obj = bin, scale = "class")

}
\seealso{
Other generic metrics: \code{\link{mAdjacency}},
  \code{\link{mArea}}, \code{\link{mNumber}},
  \code{\link{mValues}}
}
\concept{generic metrics}
