% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_derived.R
\docType{data}
\name{mPLA}
\alias{mPLA}
\title{Derived metric: Percentage of Like Adjacencies}
\format{A list of 3 elements:
\describe{
  \item{g_ii}{like adjacencies based on doble count: list(operator = "mAdjacency")}
  \item{g_ik}{sum of paired adjacencies based on double count: list(operator = "mAdjacency", type = "pairedSum")}
  \item{equation}{"(g_ii / g_ik) * 100"}
}}
\usage{
mPLA
}
\description{
Percent of the adjacencies per class that are like adjacencies.
}
\keyword{datasets}
