% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.operators.R
\name{nlmHeightmap}
\alias{nlmHeightmap}
\title{Build a landscape model based on a heightmap}
\usage{
nlmHeightmap(mat, type, fracdim, seed)
}
\arguments{
\item{mat}{bla}

\item{type}{the model used to create the heightmap. Possible options are
\code{diamondSquare} (default), ...}

\item{fracdim}{bla}

\item{seed}{bla}
}
\description{
A heightmap is the two-dimensional representation of a three-dimensional
surface, where the value of the heightmap represents the height in the
three-dimensional surface (like a digitial elevation model).
}
