% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.operators.R
\name{nlmRandom}
\alias{nlmRandom}
\title{Build a random landscape model}
\usage{
nlmRandom(mat, pattern, seed)
}
\arguments{
\item{mat}{a matrix in which the landscape model is built. If
\code{nlmRandom} is used in conjunction with \code{\link{generate}},
\code{mat} cannot be specified, because \code{generate} assigns it.}

\item{pattern}{the pattern to which random values are assigned, either
\code{"cell"}, \code{"rectangle"}, \code{"circle"}, \code{"voronoi"} or
\code{"cluster"}; see Details.}

\item{seed}{specify a seed for the random creation of numbers.}
}
\description{
Random values can be applied to different patterns within the landscape
model.
}
\details{
\itemize{ \item For \code{base = "cell"}, each cell of the resulting
  LM is assigned a random value. \item For \code{base = "rectangle"}, \item
  For \code{base = "circle"}, \item For \code{base = "voronoi"}, Gaucherel
  (2008) \item For \code{base = "cluster"}, Saura & Martínez-Millán (2000) }
}
\references{
Gaucherel, C. (2008) Neutral models for polygonal landscapes with
  linear networks. Ecological Modelling, 219, 39 - 48. Saura, S.,
  Martínez-Millán, J. (2000) Landscape patterns simulation with a modified
  random clusters method. Landscape Ecology 15, 661–678.
}
