% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain.R
\name{obtain}
\alias{obtain}
\title{Obtain spatial datasets}
\usage{
obtain(data, mask)
}
\arguments{
\item{data}{[\code{list(.)}]\cr list of \code{operators}, in which the
functions to load spatial datasets are specified. Each \code{operator} is a
list iteself and includes the operator name and its arguments as
sub-elements; see Examples.}

\item{mask}{spatial object which serves to mask the obtained.}
}
\value{
Depends on the method called, but it is always organised in lists.
  The first hierarchical level of the list contains all the spatial
  units/masks. The second level contains an entry for each dataset and its
  temporal extent. The third level contains the resulting output. Mostly this
  would be an extracted raster according to what has been specified in the
  arguments, but for some operators the output is a \code{data.frame} or
  \code{SpatialPointsDataFrame}.
}
\description{
\code{obtain} helps to extract information from various spatial (gridded)
datasets both from local and online resources.
}
\details{
\code{obtain} expects an individual directory for each dataset,
  where all the files of this particular dataset are located. \code{rasterTools}
  provides a list of paths \code{getOption("rtPaths")}, which contains the
  directories and urls of the local and online resources, respectively. In
  case the dataset files are not available locally and if a url is available,
  an attempt to download the files is undertaken. That means typically you do
  not have to use the dataset specific \code{download*} functions manually,
  they are documented nevertheless for those who are interested. In case an
  error occurs, it might be necessary to specify new paths in
  \code{rtPaths} (see \code{\link{updatePaths}}).

  Recently supported datasets are: \itemize{ \item \code{\link{oCLC}}: Corine
  Land Cover data \item \code{\link{oEMMA}}: Occurence data of the
  \emph{Atlas of European Mammals} \item \code{\link{oGFC}}: Global Forest
  Change dataset \item \code{\link{oMODIS}}: MODIS products \item
  \code{oSENTINEL}: }
}
\examples{
\dontrun{
require(magrittr)

# load and outline masks from file with locations
myLocations <- loadData(files = "locations.csv",
                        localPath = system.file("csv", package="rasterTools"))
myMask <- gGroup(geom = myLocations, distance = 10000) \%>\%
  geomRectangle() \%>\%
  gToSp(crs = projs$laea)# \%>\%
  gBuffer(width = 1000, byid = TRUE)

# grab the data
myData <- obtain(data = myDatasets, mask = myMask)
}
}
