% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oCLC.R
\name{oCLC}
\alias{oCLC}
\title{Obtain Corine Land Cover data}
\usage{
oCLC(mask = NULL, years = NULL)
}
\arguments{
\item{mask}{[\code{geom} | \code{SpatialPolygon*}]\cr spatial object of which
the extent is the area of interest.}

\item{years}{[\code{integerish(.)}]\cr year(s) for which CLC data should be
extracted; see Details.}
}
\value{
A \code{RasterLayer} of clc data.
}
\description{
Obtain data from the 'Corine Land Cover'
\href{http://land.copernicus.eu/pan-european/corine-land-cover}{dataset}
}
\details{
The CLC data dataset is available for the years 1990, 2000, 2006 and
  2012. It covers 45 land cover classes for many European countries. It was
  compiled by the European Environmental Agency.
}
\examples{
\dontrun{

require(magrittr)

# load and outline masks from file
myMask <- loadData(files = "aWindow.csv",
                   localPath = system.file("csv", package="rasterTools")) \%>\%
  geomRectangle() \%>\%
  setCRS(crs = projs$laea)

# extract CLC data for the derived mask
myCLC <- oCLC(mask = myMask, years = c(2006, 2012))
visualise(gridded = myCLC, trace = TRUE)

# get the (updated) bibliography
reference(style = "bibtex")
}
}
\seealso{
Other obtain operators: \code{\link{oEMMA}},
  \code{\link{oGFC}}, \code{\link{oMODIS}},
  \code{\link{oSENTINEL}}
}
