% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oEFTA.R
\name{oEFTA}
\alias{oEFTA}
\alias{downloadEFTA}
\title{Obtain data from the European Atlas of Forest Tree Species}
\usage{
oEFTA(mask = NULL, species = NULL, inclMeta = FALSE)

downloadEFTA(file = NULL, localPath = NULL)
}
\arguments{
\item{mask}{[\code{geom} | \code{SpatialPolygon*}]\cr spatial object of which
the extent is the area of interest.}

\item{species}{[\code{character(.)}]\cr name(s) of species to get data for.
Can be abbreviated if a \code{\link{catalog}} is provided.}

\item{inclMeta}{[\code{logical(1)}]\cr output merely the occurence data
(\code{FALSE}, default), or output additionally metadata on the species
(\code{TRUE})?}

\item{file}{[\code{character(1)}]\cr the name of the file to download.}

\item{localPath}{[\code{character(1)}]\cr the local path where files are
supposed to be stored (i.e. on your harddisc).}
}
\value{
bla
}
\description{
Obtain data from the European Atlas of Forest Tree Species
}
\details{
https://w3id.org/mtv/FISE/map-data-RPP/v0-3-2/internet/*
https://w3id.org/mtv/FISE/map-data-RPP/v0-3-2/internet/Acer-campestre
}
\section{Functions}{
\itemize{
\item \code{downloadEFTA}: function to download data related to the EMMA dataset
}}

\examples{

# something
}
\references{
de Rigo, D., Caudullo, G., Houston Durrant, T., San-Miguel-Ayanz,
  J., 2016. The European Atlas of Forest Tree Species: modelling, data and
  information on forest tree species. In: San-Miguel-Ayanz, J., de Rigo, D.,
  Caudullo, G., Houston Durrant, T., Mauri, A. (Eds.), European Atlas of
  Forest Tree Species. Publ. Off. EU, Luxembourg, pp.
}
\seealso{
Other obtain operators: \code{\link{oCLC}},
  \code{\link{oEMMA}}, \code{\link{oGFC}},
  \code{\link{oMODIS}}, \code{\link{oWCLIM}}
}
