% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oEMMA.R
\name{oEMMA}
\alias{oEMMA}
\alias{downloadEMMA}
\title{Obtain data from the Atlas of European Mammals.}
\usage{
oEMMA(mask = NULL, species = NULL, version = 1, ...)

downloadEMMA(file = NULL, localPath = NULL)
}
\arguments{
\item{mask}{[\code{geom} | \code{Spatial*}]\cr spatial object of which the
extent is the area of interest.}

\item{species}{[\code{character(.)}]\cr name(s) of species to get occurrence
for. Can be abbreviated if a \code{\link{catalog}} is provided.}

\item{version}{[\code{integerish(1)}]\cr The version of the \emph{Atlas of
European Mammals}; the recent version is 1, but an update is in the making.}

\item{...}{[various]\cr other arguments.}

\item{file}{[\code{character(1)}]\cr the name of the file to download.}

\item{localPath}{[\code{character(1)}]\cr the local path where files are
supposed to be stored (i.e. on your harddisc).}
}
\value{
A data-frame of the species of interest occuring in the area outlined
  by \code{mask}.
}
\description{
Obtain occurence data of mammals in Europe of the first EMMA
\href{https://www.european-mammals.org/}{dataset}.
}
\details{
The object provided in \code{mask} is treated as a single mask,
  irrespective of that object consisting of only one or several features. The
  extent comprising all features (point(s), line(s), polygon(s)) is used as
  area of interest. This is in contrast to \code{\link{obtain}}, where a mask
  may consist of several features, each of which are treated as seperate
  mask.

The website \href{http://www.european-mammals.org/}{Societas
  Europaea Mammalogica} kindly offers maps of the ocurrence of all mammals in
  Europe, which are parsed by \code{oEMMA}. See \code{\link{load_svg}} for a
  technical explanation.

  The occurrence data are recorded and stored on a grid of 50 x 50 km. This
  is the same grid as used for the
  \href{http://www.luomus.fi/en/new-grid-system-atlas-florae-europaeae}{Atlas
  Florae Europaeae}.

  The dataset \code{\link{meta_emma}} lists all available species.
}
\section{Functions}{
\itemize{
\item \code{downloadEMMA}: function to download data related to the EMMA dataset
}}

\examples{
\dontrun{

require(magrittr)

mySpecies <- oEMMA(mask = rtGeoms$mask,
                   species = c("Apodemus agrarius",
                               "Apodemus flavicollis",
                               "Vulpes vulpes"))

# Subset from a large set of files using an index
abbr_species <- function(x){
  tolower(paste(substr(unlist(strsplit(x, ' ')), 1, 3), collapse = '_'))
}
mySpecies <- catalog(path = rtPaths$emma$local,
                   type = 'svg', abbr = abbr_species) \%>\%
  subset(abbr \%in\% c("apo_agr", "apo_fla", "vul_vul")) \%$\%
  oEMMA(mask = myMask, species = original)

}
}
\references{
Mitchell-Jones A, Amori G, Bogdanowicz W, Kryštufek B, Reijnders
  P, Spitzenberger F, Stubbe M, Thissen J, Vohralík V and Zima J (1999). The
  Atlas of European Mammals. Academic Press, London
}
\seealso{
Other obtain operators: \code{\link{oCLC}},
  \code{\link{oEFTA}}, \code{\link{oGFC}},
  \code{\link{oMODIS}}, \code{\link{oWCLIM}}
}
\concept{obtain operators}
