% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oGFC.R
\name{oGFC}
\alias{oGFC}
\alias{downloadGFC}
\title{Obtain Global Forest Change data}
\usage{
oGFC(mask = NULL, years = NULL, keepRaw = FALSE)

downloadGFC(file = NULL, localPath = NULL)
}
\arguments{
\item{mask}{[\code{geom} | \code{SpatialPolygon*}]\cr spatial object of which
the extent is the area of interest.}

\item{years}{[\code{integerish(.)}]\cr year(s) for which GFC data should be
extracted; see Details.}

\item{keepRaw}{[\code{logical(1)}]\cr should the raw data be retained
(\code{TRUE}), or should only the derived data be returned (\code{FALSE},
default)?}

\item{file}{[\code{character(1)}]\cr the name of the file to download.}

\item{localPath}{[\code{character(1)}]\cr the local path where files are
supposed to be stored (i.e. on your harddisc).}
}
\value{
A \code{RasterStack} of gfc data.
}
\description{
Obtain data from the 'Global Forest Change'
\href{https://earthenginepartners.appspot.com/science-2013-global-forest}{dataset}
(\href{https://doi.org/10.1126/science.1244693}{paper})
}
\details{
The GFC dataset is based on a time-series analysis of Landsat images
  characterizing forest extent and change.

  A problem with the GFC dataset is that the gain-layer is calculated for the
  overall period from 2000 to 2014, while the loss-layer contains the loss
  events on a yearly basis. Hence, to find the true value per year for a
  raster-cell may not be straightforward. In \code{oGFC} a yearly value is
  derived by removing all loss events up to the year in question from the
  \emph{year 2000}-layer and subsequently adding the \emph{gain}-layer. Gain
  events are in nature rather diffuse and happen progressively and relatively
  slowely throughout time. A raster cell, which was marked as "forest absent"
  in 2000 and which had a positive gain-value by 2014, "became tree" within
  this time-frame (i.e. became over 5 m tall). This event, "becoming tree",
  presumably does not mean that the vegetation in this raster cell grew to a
  tall and mature forest. Much rather it would have some hight between 5 m
  and what a tree in the given region can grow in that short time. To get a
  more accurate estimation of the forest cover - particularly in very dynamic
  landscapes - it might be wise to weigh the forest cover with some sort of
  productivity and/or macroclimate dataset, because more suitable sites
  result in faster growth of trees.
}
\section{Functions}{
\itemize{
\item \code{downloadGFC}: function to download data related to the GFC dataset
}}

\examples{
\dontrun{

require(magrittr)

# load and outline masks from file
myMask <- loadData(files = "aWindow.csv",
                   localPath = system.file("csv", package="rasterTools")) \%>\%
  geomRectangle() \%>\%
  setCRS(crs = projs$laea)

# extract GFC data for the derived mask
myGFC <- oGFC(mask = myMask,
              years = c(2002, 2006, 2010, 2014))
visualise(gridded = myGFC$treecover_2002, trace = TRUE)

# get the (updated) bibliography
reference(style = "bibtex")
}
}
\references{
Hansen, M.C., Potapov, P.V., Moore, R., Hancher, M., Turubanova,
  S.A., Tyukavina, A., Thau, D., Stehman, S.V., Goetz, S.J., Loveland, T.R.,
  Kommareddy, A., Egorov, A., Chini, L., Justice, C.O., Townshend, J.R.G.,
  2013. High-Resolution Global Maps of 21st-Century Forest Cover Change.
  Science 342, 846–850.
}
\seealso{
Other obtain operators: \code{\link{oCLC}},
  \code{\link{oEMMA}}, \code{\link{oMODIS}}
}
