% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oSENTINEL.R
\name{oSENTINEL}
\alias{oSENTINEL}
\title{Obtain data from the Sentinel dataset}
\usage{
oSENTINEL(mask, period, band, index, path = ".")
}
\arguments{
\item{mask}{spatial object of the area of interest of the spatial dataset.}

\item{period}{year of interest (YYYY), optionally concatenated to day of the
year (DDD)(see Details).}

\item{band}{band(s) of interest. Can be numerical or a character string of
the its name.}

\item{index}{look-up-table of abbreviations with which tile names match
(built with \code{index}).}

\item{path}{exact path to the directory where the dataset files are located.}
}
\description{
Obtain data from the Sentinel dataset
}
\examples{
\dontrun{ # first set 'options(myPath = ...)' and download files
## search for the exact path, where the dataset has been stored
path_gfc <- listPaths(path = getOption("myPath"), pattern = c("(?i)sentinel$"),
                      full_names = TRUE, recursive = TRUE)
}
}
\seealso{
Other obtain operators: \code{\link{oCLC}},
  \code{\link{oEMMA}}, \code{\link{oGFC}},
  \code{\link{oMODIS}}
}
