% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oWCLIM.R
\name{oWCLIM}
\alias{oWCLIM}
\alias{downloadWCLIM}
\title{Obtain global climate data}
\usage{
oWCLIM(mask = NULL, variable = NULL, month = c(1:12),
  resolution = 0.5, ...)

downloadWCLIM(file = NULL, localPath = NULL)
}
\arguments{
\item{mask}{[\code{geom} | \code{Spatial*}]\cr spatial object of which the
extent is the area of interest.}

\item{variable}{[\code{character(.)}]\cr the climatic variable of interest;
see Details.}

\item{month}{[\code{integerish(.)}]\cr the month(s) for which the data should
be extracted.}

\item{resolution}{[\code{numeric(1)}]\cr the spatial resolution in
arc-minutes.}

\item{...}{[various]\cr other arguments.}

\item{file}{[\code{character(1)}]\cr the name of the file to download.}

\item{localPath}{[\code{character(1)}]\cr the local path where files are
supposed to be stored (i.e. on your harddisc).}
}
\description{
Obtain data from the 'worldclim 2' \href{http://worldclim.org/}{dataset}
(\href{https://doi.org/10.1002/joc.5086}{paper}).
}
\details{
The object provided in \code{mask} is treated as a single mask,
  irrespective of that object consisting of only one or several features. The
  extent comprising all features (point(s), line(s), polygon(s)) is used as
  area of interest. This is in contrast to \code{\link{obtain}}, where a mask
  may consist of several features, each of which are treated as seperate
  mask.

The following variables are included:\itemize{ 
  \item \code{"tmin"}: average monthly minimum temperature [°C * 10], 
  \item \code{"tmax"}: average monthly maximum temperature [°C * 10], 
  \item \code{"tavg"}: average monthly mean temperature [°C * 10], 
  \item \code{"prec"}: average monthly precipitation [mm], 
  \item \code{"srad"}: average monthly solar radiation
  [kJ m-2 day-1], 
  \item \code{"wind"}: average monthly wind speed [m s-1],
  \item \code{"vapr"}: average monthly water vapour pressure [kPa],
  \item \code{"bio"}: all bioclimatic variables,
  \item \code{"bio_01"}: annual mean temperature,
  \item \code{"bio_02"}: mean diurnal range,
  \item \code{"bio_03"}: Isothermality,
  \item \code{"bio_04"}: temperature seasonality,
  \item \code{"bio_05"}: Max temperature of warmest month,
  \item \code{"bio_06"}: Min temperature of coldest month,
  \item \code{"bio_07"}: temperature annual Range,
  \item \code{"bio_08"}: mean temperature of wettest quarter,
  \item \code{"bio_09"}: mean temperature of driest quarter,
  \item \code{"bio_10"}: mean temperature of warmest quarter,
  \item \code{"bio_11"}: mean temperature of coldest quarter,
  \item \code{"bio_12"}: annual precipitation,
  \item \code{"bio_13"}: precipitation of wettest month,
  \item \code{"bio_14"}: precipitation of driest month,
  \item \code{"bio_15"}: precipitation seasonality,
  \item \code{"bio_16"}: precipitation of wettest quarter,
  \item \code{"bio_17"}: precipitation of driest quarter,
  \item \code{"bio_18"}: precipitation of warmest quarter,
  \item \code{"bio_19"}: precipitation of coldest quarter.}
}
\section{Functions}{
\itemize{
\item \code{downloadWCLIM}: function to download global climte datasets
}}

\examples{
\dontrun{

myWCLIM <- oWCLIM(mask = rtGeoms$mask, variable = c("tavg"),
                  month = c(5:9))
visualise(raster = myWCLIM, trace = TRUE)

# get the (updated) bibliography
reference(style = "bibtex")
}
}
\seealso{
Other obtain operators (Global): \code{\link{oESALC}},
  \code{\link{oGFC}}, \code{\link{oMODIS}}
}
\concept{obtain operators (Global)}
