% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.operators.R
\name{rBinarise}
\alias{rBinarise}
\title{Binarise the values in a raster}
\usage{
rBinarise(obj, thresh = NULL, match = NULL)
}
\arguments{
\item{obj}{[\code{RasterLayer(1)}]\cr The object to modify.}

\item{thresh}{[\code{integerish(1)}]\cr Value above which the cell will be set
to 1, below which it will be set to 0.}

\item{match}{[\code{integerish(.)}]\cr One or more values which will be set to
1, while the remaining values will be set to 0.}
}
\value{
A binary \code{RasterLayer} of the same dimension as \code{obj}.
}
\description{
Transform a raster to binary form.
}
\details{
A binary object contains only values 1 (foreground) and 0
  (background).
}
\examples{
input <- rtData$continuous
visualise(rBinarise(input, thresh = 30))

visualise(rBinarise(input, match = 1), new = TRUE)
}
\seealso{
Other operators to modify cell values: \code{\link{rCategorise}},
  \code{\link{rDistance}}, \code{\link{rFillNA}},
  \code{\link{rOffset}}, \code{\link{rPermute}},
  \code{\link{rRange}}, \code{\link{rSubstitute}}
}
