% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.operators.R
\name{rBlend}
\alias{rBlend}
\title{Blend two rasters with each other}
\usage{
rBlend(obj, overlay, patches = FALSE, weight = 1)
}
\arguments{
\item{obj}{[\code{RasterLayer(1)}]\cr The object to modify.}

\item{overlay}{[\code{RasterLayer(1)} | \code{matrix(1)}]\cr the raster that
should be blended with the primary raster (obj); has to have the same
dimension as the primary raster.}

\item{patches}{[\code{logical(1)}]\cr should the blend be carried out while
considering foreground patches in the primary raster (\code{TRUE}), or
should the blend be carried out without this consideration (\code{FALSE},
default).}

\item{weight}{[\code{numeric(1)}]\cr the weight the overlay raster should
have as multiple (or fraction) of the primary raster.}
}
\value{
A \code{RasterLayer} of the same dimension as \code{primary}, in
  which an overlay has been blended with the primary raster.
}
\description{
Blend two rasters simply by adding (and weighing) their values per cell.
}
\details{
In case \code{patches = TRUE}, the cells in the overlay are grouped
  according to the foreground patches in the primary raster and all cells of
  this group are assigned their common average, then the blend is carried
  out.

  A blend operation currently is defined as 'overlay*weight + obj'. To
  aggregate more than two \code{RasterLayers} with arbitrary functions, see
  \code{\link{rReduce}}.
}
\examples{
# define primary raster ...
input <- rtData$continuous
patches <- rPatches(rBinarise(input, thresh = 30))

# ... and an overlay from a matrix
m <- matrix(nrow = 56, ncol = 60, data = 0)
m[c(5:25), c(5:50)] <- 10
mask <- raster::raster(m, xmn=0, xmx=60, ymn=0, ymx=56, crs=NA)

# blend while considering patches in the primary raster
blended <- rBlend(patches, overlay = mask, patches = TRUE)
visualise(raster::stack(patches, blended))

# blend while not considering patches
visualise(rBlend(patches, overlay = mask, weight = 10))
}
\seealso{
Other operators to modify a raster: \code{\link{rReduce}},
  \code{\link{rRescale}}, \code{\link{rSegregate}}
}
\concept{operators to modify a raster}
