% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.operators.R
\name{rBounded}
\alias{rBounded}
\title{Select cells with values between an upper and lower threshold in a raster}
\usage{
rBounded(obj, range = NULL, background = NULL)
}
\arguments{
\item{range}{[\code{numeric(2)}]\cr minimum and maximum value between which
the values will be selected.}

\item{background}{[\code{integerish(1)}]\cr the value any cell with value NA
should have.}

\item{ras}{[\code{RasterLayer(1)}]\cr The object to modify.}
}
\value{
A \code{RasterLayer} of the same dimension as \code{obj}, where the
  cells with values covered by \code{range} retain their value and other
  cells are set to \code{NA}.
}
\description{
Select cells with values between an upper and lower threshold in a raster
}
\examples{
input <- rtData$continuous
visualise(rBounded(input, range = c(40, 80)))
}
\seealso{
Other operators to select a subset of cells: \code{\link{rGreater}},
  \code{\link{rLess}}, \code{\link{rMask}},
  \code{\link{rMatch}}
}
