% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.operators.R
\name{rCentroid}
\alias{rCentroid}
\title{Determine the centroid of patches in a raster}
\usage{
rCentroid(obj, output = "raster", background = NULL)
}
\arguments{
\item{obj}{[\code{RasterLayer(1)}]\cr The object to modify.}

\item{output}{[\code{character(1)}]\cr the type in which the output should be
returned. Either \code{"raster"} (default) or \code{"geom"}.}

\item{background}{[\code{integerish(1)}]\cr the value any cell with value NA
should have.}
}
\value{
Depending on \code{output}, either \enumerate{ \item a
  \code{RasterLayer} of the same dimension as \code{obj}, in which the
  centroid of each foreground patch has the value of the patch-number
  identified with \code{\link{rPatches}} and where the remaining cells of
  each foreground patch have the value NA or \item a geometry object
  of the centroids.}
}
\description{
The centroid is the average location of all cells of a foreground patch.
}
\examples{
input <- rtData$continuous
patches <- rPatches(rBinarise(input, thresh = 30))
visualise(rCentroid(patches))

# get centroid coordinates
centroid_coords <- rCentroid(patches, output = "geom")
visualise(gridded = patches, geom = centroid_coords, col = "red")
}
\seealso{
Other operators to determine objects: \code{\link{rPatches}},
  \code{\link{rSkeletonise}}
}
