% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_modis}
\alias{ref_modis}
\title{References of the MODIS dataset}
\format{The \code{list} with 12 elements contains the \code{bibentry} objects
  that are needed to include a reference for the supported MODIS datasets.}
\usage{
ref_modis
}
\description{
References of the MODIS dataset
}
\keyword{datasets}
