% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{reference}
\alias{reference}
\title{List of references}
\usage{
reference(outFile = NULL, print = TRUE, style = "bibtex", ...)
}
\arguments{
\item{outFile}{[\code{character(1)}]\cr an optional file to which the
bibliography should be saved.}

\item{print}{[\code{logical(1)}]\cr should the output be printed in the
console (\code{TRUE}, defailt) or should it merely be returned
(\code{FALSE})?}

\item{style}{[\code{character(1)}]\cr the print style (by default
\code{"bibtex"}). If present, must be a unique abbreviation (with case
ignored) of the available styles, see \code{\link[utils]{bibentry}}.}

\item{...}{[various]\cr additional arguments for the \code{format()} method
of \code{bibentry}.}
}
\value{
A list of all the references that result from the operations that
  were used in the recent session.
}
\description{
Put together a list of references (a bibliography) of all the used datasets
and R-packages
}
\examples{
# just return the list to the console ...
reference(style = "latex")

# ... or store it in your project path
reference(outFile = "myBib.bib")
}
