% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.operators.R
\name{rFillNA}
\alias{rFillNA}
\title{Fill NA values in a raster}
\usage{
rFillNA(obj, with = 0)
}
\arguments{
\item{with}{[\code{integerish(1)}]\cr value with which NA should be replaced.}

\item{ras}{[\code{RasterLayer(1)}]\cr The object to modify.}
}
\value{
a \code{RasterLayer} of the same dimension as \code{obj}, where value
  \code{NA} has been replaced by value \code{with}.
}
\description{
Assign a value to cells with NA
}
\examples{
input <- rtData$continuous
patches <- rPatches(rBinarise(input, thresh = 30))

visualise(raster::stack(patches, rFillNA(patches)))
}
\seealso{
Other operators to modify cell values: \code{\link{rBinarise}},
  \code{\link{rCategorise}}, \code{\link{rDistance}},
  \code{\link{rOffset}}, \code{\link{rPermute}},
  \code{\link{rRange}}, \code{\link{rSubstitute}}
}
