% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.operators.R
\name{rOffset}
\alias{rOffset}
\title{Offset the values in a raster.}
\usage{
rOffset(obj, value = 1)
}
\arguments{
\item{obj}{[\code{RasterLayer(1)}]\cr The object to modify.}

\item{value}{[\code{integerish(1)}]\cr the value by which to offset all
values in the raster (can also be negative).}
}
\value{
A \code{RasterLayer} of the same dimension as \code{obj}, in which
  all values have been offsetted by \code{value}.
}
\description{
Offset the values in a raster.
}
\examples{
input <- rtData$continuous
binarised <- rBinarise(input, thresh = 30)
visualise(rOffset(binarised, value = 2))
}
\seealso{
Other operators to modify cell values: \code{\link{rBinarise}},
  \code{\link{rCategorise}}, \code{\link{rDistance}},
  \code{\link{rFillNA}}, \code{\link{rPermute}},
  \code{\link{rRange}}, \code{\link{rSubstitute}}
}
