% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.operators.R
\name{rPermute}
\alias{rPermute}
\title{Apply a permutation to the cell values of a raster}
\usage{
rPermute(obj, type = "invert", by = NULL)
}
\arguments{
\item{obj}{[\code{RasterLayer(1)}]\cr The object to modify.}

\item{type}{[\code{character(1)}]\cr the type of permutation that should be
applied. Either \code{"invert"}, \code{"revert"}, \code{"descending"},
\code{"ascending"} or \code{"cycle"}.}

\item{by}{[\code{integerish(1)}]\cr value by which to apply the permutation;
recently only for \code{type = "cycle"}.}
}
\value{
A \code{RasterLayer} of the same dimensions as \code{obj}, in which
  all values have been permuted according to the chosen \code{type}.
}
\description{
The permutation of a set of cell values leads to a systematic (re)arrangement
of all the members of the set.
}
\examples{
input <- rtData$continuous

# invert background/foreground
visualise(raster::stack(input, rPermute(input)))

# sort patch values ascending
patches <- rPatches(rBinarise(input, thresh = 30))
visualise(raster::stack(patches, rPermute(patches, type = "ascending")))

# cycle values backwards by 30
visualise(raster::stack(input, rPermute(input, type = "cycle", by = -30)))
}
\seealso{
Other operators to modify cell values: \code{\link{rBinarise}},
  \code{\link{rCategorise}}, \code{\link{rDistance}},
  \code{\link{rFillNA}}, \code{\link{rOffset}},
  \code{\link{rRange}}, \code{\link{rSubstitute}}
}
\concept{operators to modify cell values}
