% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.operators.R
\name{rRange}
\alias{rRange}
\title{Rescale the range of values of a raster}
\usage{
rRange(obj, range = NULL)
}
\arguments{
\item{obj}{[\code{RasterLayer(1)}]\cr The object to modify.}

\item{range}{[\code{integer(2)}]\cr vector of minimum and maximum value to
which the values will be scaled.}
}
\value{
A \code{RasterLayer} of the same dimension as \code{obj}, where the
  cell values have been rescaled.
}
\description{
Rescale the range of values of a raster
}
\examples{
input <- rtData$continuous
scaled <- rRange(input, range = c(0, 25))
visualise(raster::stack(input, scaled))
}
\seealso{
Other operators to modify cell values: \code{\link{rBinarise}},
  \code{\link{rCategorise}}, \code{\link{rDistance}},
  \code{\link{rFillNA}}, \code{\link{rOffset}},
  \code{\link{rPermute}}, \code{\link{rSubstitute}}
}
\concept{operators to modify cell values}
