% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.operators.R
\name{rReduce}
\alias{rReduce}
\title{Combine a raster stack after segregation}
\usage{
rReduce(obj, by = NULL, fun = function(x) sum(x, na.rm = TRUE),
  weights = NULL, direction = "right")
}
\arguments{
\item{obj}{[\code{RasterLayer(1)}]\cr The object to modify.}

\item{by}{[\code{list(.)}]\cr the cell values by which the layers shall be
aggregated; see Details.}

\item{fun}{[\code{function(1)}]\cr find the resulting value of the combined
layers.}

\item{weights}{[\code{numeric(length(obj))}]\cr weight by which the values in
each layer are multiplied before applying \code{fun}.}

\item{direction}{[\code{character(1)}]\cr the direction into which the
\code{RasterLayer}s of \code{obj} are supposed to be combine. Either
\code{"right"} (default) or \code{"left"}; see Details.}
}
\value{
a \code{RasterLayer} or \code{RasterStack} of the same dimensions as
  \code{obj}, in which the layers of \code{obj} are aggregated into a smaller
  set of layers.
}
\description{
Successively combine the layers of a raster stack with each other (similar to
\code{\link{Reduce}}).
}
\details{
The argument \code{direction} takes the direction into which the
  layers should be combined. \code{"right"} means that layers are combined
  from left to right. \code{rReduce} is based on the functional
  \code{\link{Reduce}}, where this wording is handled differently.

  The number of layers in the aggregated raster depends on the length of the
  list in \code{by}. If by is left empty, everything is written into one
  \code{RasterLayer} object. Values to be aggregated in a \code{RasterLayer}
  are in the same list element.
}
\examples{
input <- rtData$continuous
patches <- rPatches(rBinarise(input, thresh = 30))
myPatches <- rSegregate(patches)

# revert the segregation
visualise(rReduce(myPatches))

# group patches
by <- list(c(1:14), c(15:28))
visualise(rReduce(myPatches, by = by))

# select a subset of patches
by <- list(c(1, 3, 5, 7, 9))
visualise(rReduce(myPatches, by = by))
}
\seealso{
Other operators to modify a raster: \code{\link{rBlend}},
  \code{\link{rRescale}}, \code{\link{rSegregate}}
}
\concept{operators to modify a raster}
