% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.operators.R
\name{rSkeletonise}
\alias{rSkeletonise}
\title{Determine the skeleton of foreground patches in a raster}
\usage{
rSkeletonise(obj, method = "hitormiss", kernel = NULL, background = NULL)
}
\arguments{
\item{obj}{[\code{RasterLayer(1)}]\cr The object to modify.}

\item{method}{[\code{character(1)}]\cr the method to determine the skeleton.
Either \code{"hitormiss"} (default), \code{"lantuejoul"} or
\code{"beucher"}.}

\item{kernel}{[\code{matrix(1)}]\cr scan the raster with this kernel (default
is a 3 by 3 cells diamond kernel).}

\item{background}{[\code{integerish(1)}]\cr the value any cell with value NA
should have.}
}
\value{
a \code{RasterLayer} of the same dimensions as \code{obj}, in which
  foreground patches have been transformed into their morphological
  skeletons.
}
\description{
The morphological skeleton of a patch is a binary skeletal remnant that
preserves the extent and connectivity (i.e. the topology) of the patch.
}
\details{
For details, refer to \code{\link[mmand]{skeletonise}} of which
  \code{rSkeletonise} is a wrapper.
}
\examples{
input <- rtData$continuous
binarised <- rBinarise(input, thresh = 30)

visualise(raster::stack(binarised, rSkeletonise(binarised)))
}
\seealso{
\code{\link{rDistance}}

Other operators to determine objects: \code{\link{rCentroid}},
  \code{\link{rPatches}}
}
