% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.operators.R
\name{rSubstitute}
\alias{rSubstitute}
\title{Substitute values in a raster}
\usage{
rSubstitute(obj, old = NULL, new = NULL)
}
\arguments{
\item{obj}{[\code{RasterLayer(1)}]\cr The object to modify.}

\item{old}{[\code{integerish(.)}]\cr values to be substituted.}

\item{new}{[\code{integerish(.)}]\cr  values to substitute with.}
}
\value{
A \code{RasterLayer} of the same dimensions as \code{obj}, in which
  \code{old} values have been replaced by \code{new} values.
}
\description{
Substitute values in a raster
}
\examples{
input <- rtData$categorial
substituted <- rSubstitute(input, old = c(41:47), new = 40)
visualise(raster::brick(input, substituted))
}
\seealso{
Other operators to modify cell values: \code{\link{rBinarise}},
  \code{\link{rCategorise}}, \code{\link{rDistance}},
  \code{\link{rFillNA}}, \code{\link{rOffset}},
  \code{\link{rPermute}}, \code{\link{rRange}}
}
