% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rtGeoms}
\alias{rtGeoms}
\title{Two example Geoms}
\format{The list contains two objects of class geom, a polygon and a couple
  of points. Both are mostly used in the example and test-sections of the
  rasterTools package.}
\usage{
rtGeoms
}
\description{
A set of two geometries.
}
\keyword{datasets}
