% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rtKernels}
\alias{rtKernels}
\title{A set of 2D kernels}
\format{The list has the 4 elements 'disc', 'diamond', 'square' and
  'rectangle', where each element contains a set of differently sized
  matrixes that contain values 0, 1 or NA.}
\usage{
rtKernels
}
\description{
This objects stores several kernels that can be used with various
morphological operators.
}
\keyword{datasets}
