% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rtMD5}
\alias{rtMD5}
\title{MD5 checksums}
\format{The tibble has the 4 columns 'dataseries', 'file', 'md5' and 'size', where
  each file has been assigned the MD5 checksum that is used to validate the
  correct download of this file, and the expected size in bytes.}
\usage{
rtMD5
}
\description{
MD5 checksums of various files that are handled by rasterTools
}
\keyword{datasets}
