% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rtPalette}
\alias{rtPalette}
\title{Create non-linear palettes}
\usage{
rtPalette(colors, steps = NULL, ...)
}
\arguments{
\item{colors}{[\code{character(.)}]\cr colours to interpolate. must be a valid
argument to \code{\link[grDevices]{col2rgb}}.}

\item{steps}{[\code{integerish(.)}]\cr the number of colours between each of
\code{'colors'}.}

\item{...}{arguments to pass to \code{colorRampPalette}.}
}
\description{
A wrapper around \code{colorRampPalette} to create palettes where the number
of interpolated colours is not linear, i.e. there may be a different number
of colour values between the colors that shall be interpolated.
}
\examples{
myColours <- terrain.colors(5)

linearPalette <- rtPalette(colors = myColours)
pie(x = rep(1, 50), col = linearPalette(50))

nonlinearPalette <- rtPalette(colors = myColours, steps = c(20, 20, 2, 2))
pie(x = rep(1, 50), col = nonlinearPalette(50))
}
