% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rtPaths}
\alias{rtPaths}
\title{Dataset paths}
\format{A list with as many elements as there are datasets currently defined.
  Each list has the elements 'local', where the path to the local storage of
  the dataset is given, and 'online', where the dataset can be donwloaded.}
\usage{
rtPaths
}
\description{
All the paths to the datasets that are of interest.
}
\keyword{datasets}
