% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{rtTheme-class}
\alias{rtTheme-class}
\alias{themeClass}
\title{Theme class and methods}
\description{
Theme class and methods
}
\section{Slots}{

\describe{
\item{\code{title}}{[\code{list(3)}]\cr}

\item{\code{box}}{[\code{list(4)}]\cr}

\item{\code{xAxis}}{[\code{list(5)}]\cr}

\item{\code{yAxis}}{[\code{list(5)}]\cr}

\item{\code{grid}}{[\code{list(5)}]\cr}

\item{\code{legend}}{[\code{list(10)}]\cr}

\item{\code{geom}}{[\code{list(7)}]\cr}

\item{\code{raster}}{[\code{list(2)}]\cr}
}}

