% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rtTheme}
\alias{rtTheme}
\title{Default visualising theme}
\format{An object of class \code{rtTheme} of length 1.}
\usage{
rtTheme
}
\description{
Default visualising theme
}
\keyword{datasets}
