% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{scaleVals}
\alias{scaleVals}
\title{Scale values}
\usage{
scaleVals(mat, range)
}
\arguments{
\item{mat}{[\code{matrix(1)}]\cr the matrix, in which values shall be scaled
to \code{range}.}

\item{range}{[\code{integerish(2)}]\cr the range, to which values in
\code{mat} should be scaled to.}
}
\value{
a matrix of scaled values
}
\description{
Scale the values of a matrix to a specific range
}
