% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.R
\docType{methods}
\name{setCRS}
\alias{setCRS}
\alias{setCRS,geom-method}
\alias{setCRS,Raster-method}
\alias{setCRS,Spatial-method}
\alias{setCRS,sf-method}
\title{Set (or transform) the coordinate reference system (generic)}
\usage{
setCRS(x, crs, ...)

\S4method{setCRS}{geom}(x, crs)

\S4method{setCRS}{Raster}(x, crs, ...)

\S4method{setCRS}{Spatial}(x, crs)

\S4method{setCRS}{sf}(x, crs)
}
\arguments{
\item{x}{the object for which to set the coordinate reference system.}

\item{crs}{[\code{character(1)}]\cr the coordinate reference system to set
for this object.}

\item{...}{other arguments.}
}
\description{
In case an object has not yet assigned a coordinate reference system, this
function simply assigns it. In case the object has already a valid crs, a
transformation to the new crs will be carried out. The transformation is
computed with the standard defined in the \code{rgdal} package.
}
\section{Methods (by class)}{
\itemize{
\item \code{geom}: set the coordinate reference system of a \code{geom}

\item \code{Raster}: set the coordinate reference system of any \code{Raster*} object

\item \code{Spatial}: set the coordinate reference system of any \code{Spatial*} object

\item \code{sf}: set the coordinate reference system of a \code{sf} object
}}

\seealso{
\code{\link{getCRS}}
}
