% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{setParameters}
\alias{setParameters}
\title{Set the scale of a plot}
\usage{
setParameters(scale = NULL, to = NULL, theme = NULL)
}
\arguments{
\item{scale}{the property of the plot that shall be scaled}

\item{to}{the attribute to which \code{x} shall be scaled.}

\item{theme}{the theme from which the graphical parameters are taken.}
}
\value{
a list of parameters to a grob.
}
\description{
Set the scale of a plot
}
