% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPaths.R
\name{setPaths}
\alias{setPaths}
\title{Manage the internal paths of rasterTools}
\usage{
setPaths(root = NULL, project = NULL, local = NULL, remote = NULL)
}
\arguments{
\item{root}{[\code{character(1)}]\cr the root directory where all spatial
files are already organised or should be saved to.}

\item{project}{[\code{character(1)}]\cr the project directory where the
output of various functions will be stored to.}

\item{local}{[\code{list(.)}]\cr a named list pointing to the directory where
the local resources are located.}

\item{remote}{[\code{list(.)}]\cr a named list pointing to the url where the
remote resources are located.}
}
\value{
Changes entries of the internal dataset \code{rtPaths}.
}
\description{
Manage the internal paths of rasterTools
}
\details{
When running the function \code{setPaths} the internal object
  \code{\link{rtPaths}} is update. In case \code{root} is given, this
  directory is searched for folders of all the specified datasets, and their
  path are assigned in \code{rtPaths}. With \code{local} or \code{remote},
  the local or online directory of a specific dataset can be updated.
}
\examples{
\dontrun{

# run this when starting rasterTools for the first time
getOption("rtPaths")
setPaths(root = "/path/to/the/spatial/files")

# in case a dataset is at an exotic locations
setPaths(local = list(sentinel = "/path/outside/of/root"))

# change the (outdated) url of some datasets
setPaths(remote = list(sentinel = "https://new_url_to_sentinel/")

getOption("rtPaths")
}
}
