% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.R
\docType{methods}
\name{setTable}
\alias{setTable}
\alias{setTable,geom-method}
\alias{setTable,RasterLayer-method}
\title{Set the table (generic)}
\usage{
setTable(x, table, ...)

\S4method{setTable}{geom}(x, table)

\S4method{setTable}{RasterLayer}(x, table)
}
\arguments{
\item{x}{the object to which to assign \code{table}.}

\item{table}{[\code{data.frame(.)}]\cr the attribute table.}

\item{...}{other arguments.}
}
\description{
Set the table (generic)
}
\section{Methods (by class)}{
\itemize{
\item \code{geom}: set the attribute table of a \code{geom}

\item \code{RasterLayer}: set the attribute table of a \code{RasterLayer}
}}

