% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.R
\docType{methods}
\name{setWindow}
\alias{setWindow}
\alias{setWindow,geom-method}
\title{Set the reference window (generic)}
\usage{
setWindow(x, to, ...)

\S4method{setWindow}{geom}(x, to)
}
\arguments{
\item{x}{the object in which to set the reference window.}

\item{to}{[\code{data.frame(1)}]\cr two oposing corners or all four
corners of the rectangle.}

\item{...}{other arguments.}
}
\description{
Set the reference window (generic)
}
\section{Methods (by class)}{
\itemize{
\item \code{geom}: set the reference window of a \code{geom}
}}

