% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{show,rtTheme-method}
\alias{show,rtTheme-method}
\title{Print rtTheme in the console}
\usage{
\S4method{show}{rtTheme}(object)
}
\arguments{
\item{object}{[\code{rtTheme}]\cr object to \code{show}.}
}
\description{
Print rtTheme in the console
}
