% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.operators.R
\name{spmHeightmap}
\alias{spmHeightmap}
\title{Generate a heightmap}
\usage{
spmHeightmap(mat, hurst = NULL, type = "diamondSquare", startDev = 1,
  seed = NULL)
}
\arguments{
\item{mat}{[\code{matrix}]\cr object to which the spatial pattern is added.}

\item{hurst}{[\code{numeric(1)}]\cr the Hurst exponent (fBm) or its
equivalent, the roughness factor (DSa). Bounded between 0 and 1.}

\item{type}{[\code{character(1)}]\cr the model according to which the pattern
is generated, either \code{"diamondSquare"} (default), \code{"brownian"} or
\code{"gaussian"}.}

\item{startDev}{bla (for now)}

\item{seed}{bla}
}
\description{
A heightmap is the two dimensional representation of a three dimensional
surface, where the value of the cells represents the height in the three
dimensional surface (a simulated digitial elevation model).
}
\details{
If this function is used in conjunction with \code{\link{generate}},
  \code{mat} cannot be specified, because \code{generate} assigns it.

\describe{ \item{Diamond Square algorithm}{The algorithm was
  originally proposed by Fournier et al. (1982), but has been enhanced since
  then. It assigns values to an empty array (one or two dimensional).
  \emph{"The algorithm recursively subdivides the interval and generates a
  scalar value at the midpoint which is proportional to the current standard
  deviation times the scale or "roughness" factor (h). [...] h is equivalent
  to the Hurst exponent (see fractional brownian motion) and can take values
  between 0 and 1."} (Fournier et al, 1982). \cr\cr The implementation here
  computes values that are at the boundary of the two dimensional array as
  average from its three only neighbours and not from the one dimensional
  version of the algorithm (Frouier et al., 1982).}
  \item{Fractional brownian motion}{} \item{Gaussian random field}{} }
}
\examples{
mat <- matrix(nrow = 100, ncol = 100, data = 0)
myHeightmap <- spmHeightmap(mat = mat, hurst = 0.4, seed = 13531)
visualise(myHeightmap)

}
\references{
Fournier A, Fussell D, Carpenter L. Computer rendering of
  stochastic models. Communications of the ACM. 1982;25:371–384

  Palmer MW. The coexistence of species in fractal landscapes. The American
  Naturalist. 1992;139:375–397

  Travis JMJ, Dytham C. A method for simulating patterns of habitat
  availability at static and dynamic range margins. Oikos. 2004;104:410–416
}
