% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.operators.R
\name{spmRandom}
\alias{spmRandom}
\title{Generate a neutral random pattern}
\usage{
spmRandom(mat, pattern = "cell", seed = NULL)
}
\arguments{
\item{mat}{[\code{matrix}]\cr object to which the spatial pattern is added.}

\item{pattern}{the pattern to which random values are assigned, either
\code{"cell"}, \code{"rectangle"}, \code{"circle"}, \code{"voronoi"} or
\code{"cluster"}; see Details.}

\item{seed}{specify a seed for the random creation of numbers.}
}
\description{
This is largely still work in progress.
}
\details{
Random values can be applied to different patterns within the landscape
model.

\itemize{ \item For \code{pattern = "cell"}, each cell of the resulting
  model is assigned a random value. \item For \code{pattern = "rectangle"}, \item
  For \code{pattern = "circle"}, \item For \code{pattern = "voronoi"}, Gaucherel
  (2008) \item For \code{pattern = "cluster"}, Saura & Martínez-Millán (2000) }
}
\examples{
mat <- matrix(nrow = 100, ncol = 100, data = 0)
myRandomPattern <- spmRandom(mat = mat)
visualise(raster = myRandomPattern)

}
\references{
Gaucherel, C. (2008) Neutral models for polygonal landscapes with
  linear networks. Ecological Modelling, 219, 39 - 48. Saura, S.,
  Martínez-Millán, J. (2000) Landscape patterns simulation with a modified
  random clusters method. Landscape Ecology 15, 661–678.
}
