% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{theme_rt}
\alias{theme_rt}
\title{Default visualising theme}
\format{An object of class \code{list} of length 7.}
\usage{
theme_rt
}
\description{
Default visualising theme
}
\keyword{datasets}
