% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tiles_emma}
\alias{tiles_emma}
\title{Tiles of the EMMA dataset}
\format{The \code{geom} with 5579 polygons consisting of a total of 28409
  vertices represents the tiles that are needed to subset the Atlas of
  European Mammals.}
\usage{
tiles_emma
}
\description{
Tiles of the EMMA dataset
}
\keyword{datasets}
