% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{toEnvironment}
\alias{toEnvironment}
\title{Object management}
\usage{
toEnvironment(object, name = NULL, envir = NULL)
}
\arguments{
\item{object}{[any]\cr the object to save in \code{envir}.}

\item{name}{[\code{character(1)}]\cr define the name under which
\code{object} should be saved in \code{envir}.}

\item{envir}{[\code{environment}]\cr the environment, which has to be created
before calling here, in which \code{object} should be saved.}
}
\value{
In case this functions is called with no value or \code{NULL} in
  \code{envir}, it will merely return \code{object}, otherwise it will check
  if \code{envir} exists and save the object there.
}
\description{
Helper function to many functions of this package that standardises the way
objects are stored in environments.
}
