% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatePaths.R
\name{updatePaths}
\alias{updatePaths}
\title{Manage the internal paths of rasterTools}
\usage{
updatePaths(root, project, set)
}
\arguments{
\item{root}{the root directory where all spatial files are already organised
or should be saved to.}

\item{project}{the directory where your current project is stored. If left
empty, it is assumed that it is the same directory as \code{root}.}

\item{set}{a named list pointing to the url where the online resources are
located (see Details).}
}
\value{
Changes entries of the internal dataset \code{rtPaths}.
}
\description{
Manage the internal paths of rasterTools
}
\details{
\code{rasterTools} includes an object \code{\link{rtPaths}}, which
  contains all paths that are currently relevant for \code{rasterTools} (i.e.
  where an 'obtain'-operator is defined). Each of the operators, but also
  other functions relying on paths take the paths from here.
}
\examples{
\dontrun{

# run this when starting rasterTools for the first time
getOption("rtPaths")
updatePaths(root = "/path/to/the/spatial/files",
            project = "/path/to/your/project")

# change the (outdated) url of some datasets
updatePaths(set = list(sentinel = "https://new_url_to_sentinel/",
                       gfc = "https://new_url_to_gfc/"))

getOption("rtPaths")
}
}
