% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.R
\docType{methods}
\name{getCoords}
\alias{getCoords}
\alias{getCoords,geom-method}
\alias{getCoords,Spatial-method}
\alias{getCoords,sf-method}
\title{Get the table of coordinates (generic)}
\usage{
getCoords(x, ...)

\S4method{getCoords}{geom}(x)

\S4method{getCoords}{Spatial}(x)

\S4method{getCoords}{sf}(x)
}
\arguments{
\item{x}{the object from which to extract the coordinates}

\item{...}{other arguments.}
}
\description{
Get the table of coordinates (generic)
}
\section{Methods (by class)}{
\itemize{
\item \code{geom}: get the table of coordinates of a \code{geom}

\item \code{Spatial}: get the table of coordinates of any \code{Spatial*} object

\item \code{sf}: get the table of coordinates of a \code{sf} object
}}

