% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.R
\docType{methods}
\name{getHistory}
\alias{getHistory}
\alias{getHistory,geom-method}
\title{Get the history (generic)}
\usage{
getHistory(x, ...)

\S4method{getHistory}{geom}(x)
}
\arguments{
\item{x}{the object from which to derive the history.}

\item{...}{other arguments.}
}
\description{
Get the history (generic)
}
\section{Methods (by class)}{
\itemize{
\item \code{geom}: get the history of a \code{geom}
}}

