% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.R
\docType{methods}
\name{getSubset}
\alias{getSubset}
\alias{getSubset,geom-method}
\alias{getSubset,sf-method}
\title{Get subset (generic)}
\usage{
getSubset(x, attr, coords, ...)

\S4method{getSubset}{geom}(x, attr, coords)

\S4method{getSubset}{sf}(x, attr, coords)
}
\arguments{
\item{x}{object to \code{subset}.}

\item{attr}{[\code{integerish(.)} | \code{logical(.)} |
\code{character(1)}]\cr rows of the attribute table to keep.}

\item{coords}{[\code{integerish(.)} | \code{logical(.)} |
\code{character(1)}]\cr coordinates to keep.}

\item{...}{other arguments.}
}
\description{
\code{getSubset} returns the subsetted object.
}
\section{Methods (by class)}{
\itemize{
\item \code{geom}: get a subset of the vertices of a \code{geom} based on a numeric

\item \code{sf}: get a subset of the vertices of a \code{sf} object based on a logical
}}

