% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.R
\docType{methods}
\name{getSubset}
\alias{getSubset}
\alias{getSubset,geom,numeric-method}
\alias{getSubset,geom,logical-method}
\title{Get subset (generic)}
\usage{
getSubset(x, subset, ...)

\S4method{getSubset}{geom,numeric}(x, subset)

\S4method{getSubset}{geom,logical}(x, subset)
}
\arguments{
\item{x}{[\code{geom}]\cr object to \code{subset}.}

\item{subset}{[\code{integerish(.)} | \code{logical(.)}]\cr elements or rows
to keep.}

\item{...}{other arguments.}
}
\description{
\code{getSubset} returns the subsetted \code{geom}.
}
\section{Methods (by class)}{
\itemize{
\item \code{x = geom,subset = numeric}: get a subset of the vertices of a \code{geom} based on a numeric

\item \code{x = geom,subset = logical}: get a subset of the vertices of a \code{geom} based on a logical
}}

