% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.R
\docType{methods}
\name{getTable}
\alias{getTable}
\alias{getTable,geom-method}
\alias{getTable,RasterLayer-method}
\alias{getTable,Spatial-method}
\alias{getTable,sf-method}
\title{Get the attribute table (generic)}
\usage{
getTable(x, ...)

\S4method{getTable}{geom}(x)

\S4method{getTable}{RasterLayer}(x)

\S4method{getTable}{Spatial}(x)

\S4method{getTable}{sf}(x)
}
\arguments{
\item{x}{the object from which to derive the attribute table.}

\item{...}{other arguments.}
}
\description{
Get the attribute table (generic)
}
\section{Methods (by class)}{
\itemize{
\item \code{geom}: get the attribute table of a \code{geom}

\item \code{RasterLayer}: get the attribute table of a \code{RasterLayer}

\item \code{Spatial}: get the attribute table of a \code{Spatial*} object

\item \code{sf}: get the attribute table of a \code{sf} object
}}

