% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_generic.R
\name{mValues}
\alias{mValues}
\title{Cell value distribution}
\usage{
mValues(obj, param = NULL, layer = NULL, groupBy = NULL)
}
\arguments{
\item{obj}{[\code{RasterLayer(1)}]\cr The object to measure.}

\item{param}{[\code{character(.)}]\cr parameter(s) to calculate; possible
parameters are \code{"mean"}, \code{"sum"}, \code{"number"}, \code{"sd"},
\code{"cv"}, \code{"iqr"}, \code{"min"}, \code{"median"}, \code{"max"},
\code{"quantile"}, \code{"weighted.mean"}, \code{"identity"}, \code{"all"}
or \code{"raw"}.}

\item{layer}{[\code{character(1)}]\cr in case \code{obj} has several layers,
specify here the layer for which the area of objects shall be calculated
(by default, the first layer).}

\item{groupBy}{another layer (typically part of the same raster::brick),
according which you want to divide the cells in the input layer. see
Details.}
}
\value{
A list with elements resulting from the stratification with the value
  of each specified parameter per object in \code{groupBby}.
}
\description{
Calculate distribution parameters of cell values of a raster
}
\details{
You can give a raster stack with at least two layers as \code{obj}
  and then use \code{groupBby} to 'mask' the cells, so to speak.
}
\examples{
cat <- rtData$categorial
con <- rtData$continuous
bin <- rBinarise(con, thresh = 40)
patches <- rPatches(bin)

input <- raster::stack(con, patches)
}
\seealso{
Other generic metrics: \code{\link{mAdjacency}},
  \code{\link{mArea}}, \code{\link{mNumber}},
  \code{\link{mPerimeter}}
}
