% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain.R
\name{obtain}
\alias{obtain}
\title{Obtain spatial datasets}
\usage{
obtain(data = NULL, mask = NULL)
}
\arguments{
\item{data}{[\code{list(.)}]\cr algorithm in which the operators to load
spatial datasets are specified. Each \code{operator} is a list iteself and
includes the operator name and its arguments as sub-elements; see Examples.}

\item{mask}{[\code{geom} | \code{Spatial*}]\cr spatial object. The extent of
vertices that are part of the same group is used to subset the spatial
information.}
}
\value{
A list of objects that is organised according to the stratification
  in \code{data}. The first hierarchical level of the list contains all the
  spatial units/masks. The second level contains an entry for each dataset
  and its temporal extent. The third level contains the resulting output.
  Mostly this would be an extracted raster according to what has been
  specified in the arguments, but for some operators the output is a
  \code{data.frame} or \code{SpatialPointsDataFrame}.
}
\description{
Extract information from various spatial (raster) dataseries both from local
and online resources.
}
\details{
\code{obtain} expects a root directory in which an individual
  directory for each dataset dwells, where in turn all the files of this
  particular dataset are located. \code{rasterTools} provides a list of paths
  \code{getOption("rtPaths")}, which contains the directories and urls of the
  local and online resources. In case the dataset files are not available
  locally and if a url is available, an attempt to download the files is
  undertaken. That means typically you do not have to use the dataset
  specific \code{download*} functions manually, they are documented
  nevertheless for those who are interested. In case an error occurs, it
  might be necessary to specify new paths in \code{rtPaths} (see
  \code{\link{setPaths}}).
  
  Recently supported datasets are \itemize{
  \item Global: \itemize{
     \item \code{\link{oGFC}}: Global Forest Change
     \item \code{\link{oMODIS}}: MODIS products
     \item \code{\link{oWCLIM}}: Worldclim
     \item \code{\link{oESALC}}: ESA CCI land-cover
  }
  \item European: \itemize{
     \item \code{\link{oCLC}}: Corine Land Cover
     \item \code{\link{oEMMA}}: Mammal occurence in the \emph{Atlas of European Mammals}
     \item \code{\link{oEFTA}}: Tree presence and habitat suitability in the \emph{European Atlas of Forest Tree Species}
  }
  }
}
\examples{

\dontrun{
require(magrittr)

# specify the datasets for which you want to get data
myDatasets <- list(list(operator = "oGFC", years = c(2005:2007)),
                   list(operator = "oMODIS", product = "mod17a3", period = 2006,
                        layer = 2))

# load and outline masks from a file with locations
myLocations <- loadData(files = "locations.csv",
                        localPath = system.file("test_datasets", package="rasterTools"))
myMask <- gGroup(geom = myLocations, distance = 10000) \%>\%
  geomRectangle() \%>\%
  gToSp(crs = projs$laea)

# grab the data
myData <- obtain(data = myDatasets, mask = myMask)
}
}
