% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oCLC.R
\name{oCLC}
\alias{oCLC}
\title{Obtain Corine Land Cover data}
\usage{
oCLC(mask = NULL, years = NULL)
}
\arguments{
\item{mask}{[\code{geom} | \code{Spatial*} | \code{sf}]\cr spatial object of
which the extent is the area of interest.}

\item{years}{[\code{integerish(.)}]\cr year(s) for which CLC data should be
extracted; see Details.}
}
\value{
A \code{RasterLayer} of clc data.
}
\description{
Obtain data from the 'Corine Land Cover'
\href{http://land.copernicus.eu/pan-european/corine-land-cover}{dataset}
(currently you have to download it manually).
}
\details{
The object provided in \code{mask} is treated as a single mask,
  irrespective of that object consisting of only one or several features. The
  extent comprising all features (point(s), line(s), polygon(s)) is used as
  area of interest. This is in contrast to \code{\link{obtain}}, where a mask
  may consist of several features, each of which are treated as seperate
  mask.

The CLC data dataset is available for the years 1990, 2000, 2006 and
  2012. It covers 45 land cover classes for many European countries. It was
  compiled by the European Environmental Agency.
}
\examples{
\dontrun{

myCLC <- oCLC(mask = rtGeoms$mask, years = c(2006, 2012))
visualise(gridded = myCLC, trace = TRUE)

# get the (updated) bibliography
reference(style = "bibtex")
}
}
\seealso{
Other obtain operators (Europe): \code{\link{oEFTA}},
  \code{\link{oEMMA}}
}
\concept{obtain operators (Europe)}
