% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oEFTA.R
\name{oEFTA}
\alias{oEFTA}
\alias{downloadEFTA}
\title{Obtain data from the European Atlas of Forest Tree Species}
\usage{
oEFTA(mask = NULL, species = NULL, type = "rpp")

downloadEFTA(file = NULL, localPath = NULL)
}
\arguments{
\item{mask}{[\code{geom} | \code{SpatialPolygon*}]\cr spatial object of which
the extent is the area of interest.}

\item{species}{[\code{character(.)}]\cr name(s) of species to get data for.
Can be abbreviated if a \code{\link{catalog}} is provided.}

\item{type}{[\code{character(1)}]\cr the dataset type, either \code{"rpp"}
(relative probability of presence) or \code{"mhs"} (maximum habitat
suitability).}

\item{file}{[\code{character(1)}]\cr the name of the file to download.}

\item{localPath}{[\code{character(1)}]\cr the local path where files are
supposed to be stored (i.e. on your harddisc).}
}
\value{
bla
}
\description{
Obtain presence and habitat suitability
\href{http://forest.jrc.ec.europa.eu/european-atlas-of-forest-tree-species/atlas-data-and-metadata/}{data}
from the
\href{http://forest.jrc.ec.europa.eu/european-atlas-of-forest-tree-species/}{European
Atlas of Forest Tree Species}
}
\details{
https://w3id.org/mtv/FISE/map-data-RPP/v0-3-2/internet/*
  https://w3id.org/mtv/FISE/map-data-RPP/v0-3-2/internet/Acer-campestre
  
  mention that values are integers and not "probabilities" between 0 and 1)
}
\section{Functions}{
\itemize{
\item \code{downloadEFTA}: function to download data related to the EMMA dataset
}}

\examples{
\dontrun{

require(magrittr)

# load and outline masks from file
myMask <- loadData(files = "aWindow.csv",
                   localPath = system.file("csv", package="rasterTools")) \%>\%
  geomRectangle() \%>\%
  setCRS(crs = projs$laea)

# extract data for the derived mask
myTrees <- oEFTA(mask = myMask,
                 species = c("Quercus robur", "Alnus incana", "Betula sp"))
visualise(gridded = myTrees, trace = TRUE)
}
}
\references{
de Rigo, D., Caudullo, G., Houston Durrant, T., San-Miguel-Ayanz,
  J., 2016. The European Atlas of Forest Tree Species: modelling, data and
  information on forest tree species. In: San-Miguel-Ayanz, J., de Rigo, D.,
  Caudullo, G., Houston Durrant, T., Mauri, A. (Eds.), European Atlas of
  Forest Tree Species. Publ. Off. EU, Luxembourg, pp.
}
\seealso{
Other obtain operators: \code{\link{oCLC}},
  \code{\link{oEMMA}}, \code{\link{oGFC}},
  \code{\link{oMODIS}}, \code{\link{oWCLIM}}
}
