% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oEFTA.R
\name{oEFTA}
\alias{oEFTA}
\alias{downloadEFTA}
\title{Obtain data from the European Atlas of Forest Tree Species}
\usage{
oEFTA(mask = NULL, species = NULL, type = "rpp")

downloadEFTA(file = NULL, localPath = NULL)
}
\arguments{
\item{mask}{[\code{geom} | \code{Spatial*}]\cr spatial object of which the
extent is the area of interest.}

\item{species}{[\code{character(.)}]\cr name(s) of species to get data for.
Can be abbreviated if a \code{\link{catalog}} is provided.}

\item{type}{[\code{character(1)}]\cr the dataset type, either \code{"rpp"}
(relative probability of presence) or \code{"mhs"} (maximum habitat
suitability).}

\item{file}{[\code{character(1)}]\cr the name of the file to download.}

\item{localPath}{[\code{character(1)}]\cr the local path where files are
supposed to be stored (i.e. on your harddisc).}
}
\value{
bla
}
\description{
Obtain presence and habitat suitability
\href{http://forest.jrc.ec.europa.eu/european-atlas-of-forest-tree-species/atlas-data-and-metadata/}{data}
from the
\href{http://forest.jrc.ec.europa.eu/european-atlas-of-forest-tree-species/}{European
Atlas of Forest Tree Species}
}
\details{
The object provided in \code{mask} is treated as a single mask,
  irrespective of that object consisting of only one or several features. The
  extent comprising all features (point(s), line(s), polygon(s)) is used as
  area of interest. This is in contrast to \code{\link{obtain}}, where a mask
  may consist of several features, each of which are treated as seperate
  mask.

The values in this dataset are originally stored as values in the
  range of 0 and 1 (the probability of presence). Here they are transformed
  to integer factors from 0 to 100.
}
\section{Functions}{
\itemize{
\item \code{downloadEFTA}: function to download data related to the EFTA dataset
}}

\examples{
\dontrun{

require(magrittr)

# load and outline masks from file
myMask <- loadData(files = "aWindow.csv",
                   localPath = system.file("csv", package="rasterTools")) \%>\%
  geomRectangle() \%>\%
  setCRS(crs = projs$laea)

# extract data for the derived mask
myTrees <- oEFTA(mask = myMask,
                 species = c("Quercus robur", "Alnus incana", 
                             "Pinus sylvestris", "Betula sp"))
visualise(gridded = myTrees, trace = TRUE)
}
}
\references{
de Rigo, D., Caudullo, G., Houston Durrant, T., San-Miguel-Ayanz,
  J., 2016. The European Atlas of Forest Tree Species: modelling, data and
  information on forest tree species. In: San-Miguel-Ayanz, J., de Rigo, D.,
  Caudullo, G., Houston Durrant, T., Mauri, A. (Eds.), European Atlas of
  Forest Tree Species. Publ. Off. EU, Luxembourg, pp.
}
\seealso{
Other obtain operators: \code{\link{oCLC}},
  \code{\link{oEMMA}}, \code{\link{oGFC}},
  \code{\link{oMODIS}}, \code{\link{oWCLIM}}
}
