% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oEMMA.R
\name{oEMMA}
\alias{oEMMA}
\alias{downloadEMMA}
\title{Obtain data of European mammals.}
\usage{
oEMMA(mask, species, version = 1, inclMeta = FALSE)

downloadEMMA(file = NULL, localPath = NULL, getGrids = NULL,
  keepGrids = FALSE)
}
\arguments{
\item{mask}{[\code{geom} | \code{SpatialPolygon*}]\cr spatial object of which
the extent is the area of interest.}

\item{species}{[\code{character(.)}]\cr name(s) of species to get occurrence
for. Can be abbreviated if a \code{\link{catalog}} is provided.}

\item{version}{[\code{integerish(1)}]\cr The version of the \emph{Atlas of
European Mammals}; the recent version is 1, but an update is in the making.}

\item{inclMeta}{[\code{logical(1)}]\cr output merely the occurence data
(\code{FALSE}, default), or output additionally metadata on the species
(\code{TRUE})?}

\item{file}{[\code{character(1)}]\cr the name of the file to download.}

\item{localPath}{[\code{character(1)}]\cr the local path where files are
supposed to be stored (i.e. on your harddisc).}

\item{getGrids}{[\code{character(1)}]\cr the online path to the gridfiles of
the EMMA project; stored under \code{rtPaths$emma$gridLinks}}

\item{keepGrids}{[\code{logic(1)}]\cr should the the downloaded and seperate
gridfiles be stored (\code{TRUE}), or should merely the overall European
grid be stored (\code{FALSE}, default)?}
}
\value{
A data-frame of the species of interest occuring in the area outlined
  by \code{mask}.
}
\description{
Obtain occurence data of mammals in Europe of the first EMMA
\href{https://www.european-mammals.org/}{dataset}
}
\details{
The website \href{http://www.european-mammals.org/}{Societas
  Europaea Mammalogica} kindly offers maps of the ocurrence of all mammals in
  Europe, which are parsed by \code{oEMMA}. See \code{\link{load_svg}} for a
  technical explanation.

  The occurrence data are recorded and stored on a grid of 50 x 50 km. This
  is the same grid as used for the
  \href{http://www.luomus.fi/en/new-grid-system-atlas-florae-europaeae}{Atlas
  Florae Europaeae}.

  The dataset \code{\link{meta_emma}} lists all available species.
}
\section{Functions}{
\itemize{
\item \code{downloadEMMA}: function to download data related to the EMMA dataset
}}

\examples{
\dontrun{

require(magrittr)

# load and outline masks from file
myMask <- loadData(files = "aWindow.csv",
                   localPath = system.file("csv", package="rasterTools")) \%>\%
  geomRectangle() \%>\%
  setCRS(crs = projs$laea)

mySpecies <- oEMMA(mask = myMask,
                   species = c("Apodemus agrarius",
                               "Apodemus flavicollis",
                               "Vulpes vulpes"))

# Subset from a large set of files using an index
abbr_species <- function(x){
  tolower(paste(substr(unlist(strsplit(x, ' ')), 1, 3), collapse = '_'))
}
mySpecies <- catalog(path = rtPaths$emma$local,
                   type = 'svg', abbr = abbr_species) \%>\%
  subset(abbr \%in\% c("apo_agr", "apo_fla", "vul_vul")) \%$\%
  oEMMA(mask = myMask, species = original)

}
}
\references{
Mitchell-Jones A, Amori G, Bogdanowicz W, Kryštufek B, Reijnders
  P, Spitzenberger F, Stubbe M, Thissen J, Vohralík V and Zima J (1999). The
  Atlas of European Mammals. Academic Press, London
}
\seealso{
Other obtain operators: \code{\link{oCLC}},
  \code{\link{oGFC}}, \code{\link{oMODIS}},
  \code{\link{oSENTINEL}}
}
