% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oMODIS.R
\name{oMODIS}
\alias{oMODIS}
\alias{downloadMODIS}
\title{Obtain MODIS data}
\usage{
oMODIS(mask = NULL, period = NULL, product = NULL, layer = NULL,
  process = TRUE)

downloadMODIS(file = NULL, localPath = NULL, getDates = NULL,
  getFiles = NULL)
}
\arguments{
\item{mask}{[\code{geom} | \code{Spatial*}]\cr spatial object of which the
extent is the area of interest.}

\item{period}{[\code{integerish(.)}]\cr year of interest (YYYY), optionally
concatenated to day of the year (DDD), which must be between \code{001} and
\code{366}.}

\item{product}{[\code{character(1)}]\cr name of the modis-product of
interest. Can be abbreviated if an \code{index} is provided.}

\item{layer}{[\code{character(1)} | \code{integerish(1)}]\cr layer(s) of
interest. Either the integer of the layers position or the name of the
layer.}

\item{process}{[\code{logical(1)}]\cr should the data be procesed further in
R (\code{TRUE}, default) or should the data merely be downloaded and stored
on the harddisc (\code{FALSE})?}

\item{file}{[\code{character(1)}]\cr the name of the file to download.}

\item{localPath}{[\code{character(1)}]\cr the local path where files are
supposed to be stored (i.e. on your harddisc).}

\item{getDates}{[\code{character(1)}]\cr the online path to the directory
where folders for all valid dates are stored (recently the first level
after selecting the product).}

\item{getFiles}{[\code{character(1)}]\cr the online path to the directory
where product files are stored.}
}
\value{
A \code{RasterLayer} or \code{RasterStack} of MODIS data.
}
\description{
Obtain data from various MODIS \href{https://modis.gsfc.nasa.gov/}{datasets}.
}
\details{
The object provided in \code{mask} is treated as a single mask,
  irrespective of that object consisting of only one or several features. The
  extent comprising all features (point(s), line(s), polygon(s)) is used as
  area of interest. This is in contrast to \code{\link{obtain}}, where a mask
  may consist of several features, each of which are treated as seperate
  mask.

The period can either be given as a single value if only one year is
  of interest, or as two values, the start and end date. If you do not
  specify DDD, it will be set to "001" and "366" for start and end year
  respectively.

  Supported MODIS-products - in the sense that predefined meta-data are
  available within this package -  are: \itemize{ \item{MOD09A1: Surface
  Reflectance, 8 day, 500 meter, February 2000 - present} \item{MOD09Q1:
  Surface Reflectance, 8 day, 250 meter, February 2000 - present}
  \item{MOD11A2: Land Surface Temperature/Emissivity, 8 day, 1000 meter,
  March 2000 - present} \item{MOD13Q1: Vegetation Indices, 16 day, 250 meter,
  February 2000 - present} \item{MOD13A1: Vegetation Indices, 16 day, 500
  meter, February 2000 - present} \item{MOD13A2: Vegetation Indices, 16 day,
  1000 meter, February 2000 - present} \item{MOD13A3: Vegetation Indices,
  monthly, 1000 meter, February 2000 - present} \item{MOD14A2: Thermal
  Anomalies and Fire, 8 day, 1000 meter, February 2000 - present}
  \item{MOD15A2H: Leaf Area Index/FPAR, 8 day, 500 meter, July 2002 -
  present} \item{MOD17A2H: Gross Primary Productivity, 8 day, 500 meter,
  February 2000 - present} \item{MOD17A3H: Net Primary Production, yearly,
  500 meter, 2000 - present} \item{MOD44W: Land-Water Mask, yearly, 250
  meter, 2000 - 2015} }

  Get additional information about the spatial background from
  \href{https://code.env.duke.edu/projects/mget/wiki/SinusoidalMODIS}{Marine
  Geospatial Ecology Tools}
}
\section{Functions}{
\itemize{
\item \code{downloadMODIS}: function to download data related to the MODIS products
}}

\examples{
\dontrun{

water <- oMODIS(mask = rtGeoms$mask, period = 2000,
                product = "MOD44W", layer = 1)
visualise(gridded = water, trace = TRUE)

# not specifying 'layer' results in a stack of all layers
myMODIS <- oMODIS(mask = rtGeoms$mask, period = 2012,
                  product = "MOD17A3")
visualise(gridded = myMODIS, trace = TRUE)

modEvi <- oMODIS(mask = rtGeoms$mask, period = c(2016100, 2016250),
                 product = "MOD13A3", layer = "EVI")
visualise(gridded = modEvi, trace = TRUE)

# get the (updated) bibliography
reference(style = "bibtex")

# the modis tiles
modWindow <- data.frame(x = c(-20015109.354, 20015109.354),
                        y = c(-10007554.677, 10007554.677))
tiles_modis <- geomTiles(window = modWindow, cells = c(36, 18),
                         crs = projs$sinu)
visualise(geom = tiles_modis)
}
}
\seealso{
Other obtain operators (Global): \code{\link{oESALC}},
  \code{\link{oGFC}}, \code{\link{oWCLIM}}
}
\concept{obtain operators (Global)}
