% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oWCLIM.R
\name{oWCLIM}
\alias{oWCLIM}
\alias{downloadWCLIM}
\title{Obtain global climate data}
\usage{
oWCLIM(mask = NULL, variable = NULL, month = c(1:12), resolution = 0.5,
  version = 2)

downloadWCLIM(file = NULL, localPath = NULL)
}
\arguments{
\item{mask}{[\code{geom} | \code{Spatial*}]\cr spatial object of which the
extent is the area of interest.}

\item{variable}{[\code{character(.)}]\cr the climatic variable of interest;
see Details to find which variables are available in which version.}

\item{month}{[\code{integerish(.)}]\cr the month(s) for which the data should
be extracted.}

\item{resolution}{[\code{numeric(1)}]\cr the spatial resolution in
arc-minutes.}

\item{version}{[\code{integerish(1)}]\cr the version for which you'd like to
obtain the worldclim data.}

\item{file}{[\code{character(1)}]\cr the name of the file to download.}

\item{localPath}{[\code{character(1)}]\cr the local path where files are
supposed to be stored (i.e. on your harddisc).}
}
\description{
Obtain data from the 'worldclim' \href{http://worldclim.org/}{dataset}
(\href{https://doi.org/10.1002/joc.1276}{paper1},
\href{https://doi.org/10.1002/joc.5086}{paper2}).
}
\details{
The object provided in \code{mask} is treated as a single mask,
  irrespective of that object consisting of only one or several features. The
  extent comprising all features (point(s), line(s),polygon(s)) is used as
  area of interest. This is in contrast to \code{\link{obtain}}, where a mask
  may consist of several features, each of which are treated as seperate
  mask.

\itemize{\item Version 1.4 includes the variables:\itemize{ \item
  \code{"tmin"}: average monthly minimum temperature [°C * 10], \item
  \code{"tmax"}: average monthly maximum temperature [°C * 10], \item
  \code{"tmean"}: average monthly mean temperature [°C * 10], \item
  \code{"prec"}: average monthly precipitation [mm], \item \code{"bio"}:
  bioclimatic variables derived from the above.} \item Version 2.0 includes
  the variables:\itemize{ \item \code{"tmin"}: average monthly minimum
  temperature [°C * 10], \item \code{"tmax"}: average monthly maximum
  temperature [°C * 10], \item \code{"tavg"}: average monthly mean
  temperature [°C * 10], \item \code{"prec"}: average monthly precipitation
  [mm], \item \code{"srad"}: average monthly solar radiation [kJ m-2 day-1],
  \item \code{"wind"}: average monthly wind speed [m s-1], \item
  \code{"vapr"}: average monthly water vapour pressure [kPa], \item
  \code{"bio"}: bioclimatic variables derived from the above.} }

  The files of version 1.4 are shipped in the 'bil' format. While the files
  are compressed into a zip-file, once unzipped, they take up a lot of space
  on the harddisc. Hence, when using this function for the first time,
  \code{downloadWCLIM} downloads and unzips the original file(s), transforms
  the monthly files into 'tif' format (which uses only about 1/10th of the
  space) and deletes the original 'bil' files. Beware that due to the size of
  the inital files this takes a lot (!) of time, especially for the 30
  arc-second files. However, when repeatedly using these data, so the hope,
  this procedure should make the future work with these data smoother.
}
\section{Functions}{
\itemize{
\item \code{downloadWCLIM}: function to download global climte datasets
}}

\examples{
\dontrun{

require(magrittr)

# load and outline masks from file
myMask <- loadData(files = "aWindow.csv",
                   localPath = system.file("csv", package="rasterTools")) \%>\%
  geomRectangle() \%>\%
  setCRS(crs = projs$laea)

# extract worldclim data for the derived mask
myWCLIM <- oWCLIM(mask = myMask,
                  variable = c("tmin", "tmax"),
                  month = c(5:9))
visualise(gridded = myWCLIM$, trace = TRUE)

# get the (updated) bibliography
reference(style = "bibtex")
}
}
\seealso{
Other obtain operators: \code{\link{oCLC}},
  \code{\link{oEFTA}}, \code{\link{oEMMA}},
  \code{\link{oGFC}}, \code{\link{oMODIS}}
}
