% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.operators.R
\name{rDilate}
\alias{rDilate}
\title{Morphologically dilate foreground patches in a raster}
\usage{
rDilate(obj, kernel = NULL)
}
\arguments{
\item{obj}{[\code{RasterLayer(1)}]\cr The object to modify.}

\item{kernel}{[\code{matrix(1)}]\cr scan the raster with this kernel
(default is a 3 by 3 cells diamond kernel).}
}
\value{
A \code{RasterLayer} of the same dimension as \code{obj}, where the
  boundaries of the binary raster have been enlarged by the given kernel.
}
\description{
The morphological operation 'dilate' adds cells at the boundary of a patch in
a binarised raster.
}
\details{
The cells of a kernel can have the values 0, 1 and NA. The kernel
  cells with values 0 and 1 are matched accurately in the input raster. The
  kernel cells with value NA will be ignored.
}
\examples{
input <- rtData$continuous
binarised <- rBinarise(input, thresh = 30)
visualise(rDilate(binarised))

# use another kernel
(myKernel <- matrix(1, 3, 3))
visualise(rDilate(binarised, kernel = myKernel), new = TRUE)
}
\seealso{
Other operators to morphologically modify a raster: \code{\link{rErode}}
}
