% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rtMD5}
\alias{rtMD5}
\title{MD5 checksums}
\format{The tibble has the 3 columns 'dataseries', 'file' and 'md5', where
  each file has been assigned the MD5 checksum that is used to validate the
  correct download of this file.}
\usage{
rtMD5
}
\description{
MD5 checksums of various files that are handled by rasterTools
}
\keyword{datasets}
