% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{setTheme}
\alias{setTheme}
\title{Create a new theme}
\usage{
setTheme(from = NULL, title = NULL, box = NULL, xAxis = NULL,
  yAxis = NULL, grid = NULL, legend = NULL, geom = NULL,
  raster = NULL)
}
\arguments{
\item{from}{[\code{theme}]\cr}

\item{title}{[\code{named list(.)}]\cr .}

\item{box}{[\code{named list(.)}]\cr .}

\item{xAxis}{[\code{named list(.)}]\cr .}

\item{yAxis}{[\code{named list(.)}]\cr .}

\item{grid}{[\code{named list(.)}]\cr .}

\item{legend}{[\code{named list(.)}]\cr .}

\item{geom}{[\code{named list(.)}]\cr .}

\item{raster}{[ \code{named list(.)}]\cr .}
}
\description{
This is merely a tentative workaround which I will improve with respect to
usability in the future. To see the default settings, type
\code{str(rtTheme)}.
}
\details{
In case a colourtable is defined in a raster, this overrides 'raster@scale'.
}
\examples{
input <- rtData$continuous
(myTheme <- setTheme(title = list(plot = FALSE)))

visualise(input, theme = myTheme)
}
